% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address-provider-en_GB.R
\name{AddressProvider_en_GB}
\alias{AddressProvider_en_GB}
\title{AddressProvider for English, Great Britain}
\description{
address methods for Great Britain.
}
\examples{
(z <- AddressProvider_en_GB$new())
z$locale
z$postcode()
z$street_name()
z$address()
z$city()
}
\seealso{
Other en: 
\code{\link{AddressProvider_en_NZ}},
\code{\link{ColorProvider_en_US}},
\code{\link{CompanyProvider_en_US}},
\code{\link{ElementProvider_en_US}},
\code{\link{InternetProvider_en_AU}},
\code{\link{InternetProvider_en_US}},
\code{\link{JobProvider_en_US}},
\code{\link{LoremProvider_en_US}},
\code{\link{PersonProvider_en_GB}},
\code{\link{PersonProvider_en_NZ}},
\code{\link{PersonProvider_en_US}},
\code{\link{PhoneNumberProvider_en_AU}},
\code{\link{PhoneNumberProvider_en_CA}},
\code{\link{PhoneNumberProvider_en_GB}},
\code{\link{PhoneNumberProvider_en_NZ}},
\code{\link{PhoneNumberProvider_en_US}},
\code{\link{PhoneNumberProvider_es_MX}},
\code{\link{PhoneNumberProvider_tr_TR}},
\code{\link{PhoneNumberProvider}},
\code{\link{SSNProvider_en_US}},
\code{\link{TaxonomyProvider_en_US}},
\code{\link{UserAgentProvider_en_US}},
\code{\link{UserAgentProvider}},
\code{\link{english-language}}

Other GB: 
\code{\link{PersonProvider_en_GB}},
\code{\link{PhoneNumberProvider_en_GB}},
\code{\link{PhoneNumberProvider_tr_TR}}
}
\concept{GB}
\concept{en}
\section{Super classes}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{\link[charlatan:AddressProvider]{charlatan::AddressProvider}} -> \code{AddressProvider_en_GB}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{locale}}{(character) en_GB Locale.}

\item{\code{city_prefixes}}{(character) Prefixes for cities.
Combining suffix and prefix makes a random city.}

\item{\code{city_suffixes}}{(character) suffixes for cities.}

\item{\code{building_number_formats}}{(character) Formats for building numbers}

\item{\code{street_suffixes}}{combined with street prefix creates a street.}

\item{\code{postcode_formats}}{Formats for postal codes}

\item{\code{postcode_sets}}{Specifically for en_GB, postcodes
form specific combinations.}

\item{\code{city_formats}}{formats for creating a city name.}

\item{\code{street_name_formats}}{formats for creating a street name.}

\item{\code{street_address_formats}}{formats for creating a street address.}

\item{\code{address_formats}}{formats for creating an address.}

\item{\code{secondary_address_formats}}{formats for creating secondary parts of
addresses.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AddressProvider_en_GB-address}{\code{AddressProvider_en_GB$address()}}
\item \href{#method-AddressProvider_en_GB-city}{\code{AddressProvider_en_GB$city()}}
\item \href{#method-AddressProvider_en_GB-street_name}{\code{AddressProvider_en_GB$street_name()}}
\item \href{#method-AddressProvider_en_GB-street_address}{\code{AddressProvider_en_GB$street_address()}}
\item \href{#method-AddressProvider_en_GB-postcode}{\code{AddressProvider_en_GB$postcode()}}
\item \href{#method-AddressProvider_en_GB-building_number}{\code{AddressProvider_en_GB$building_number()}}
\item \href{#method-AddressProvider_en_GB-clone}{\code{AddressProvider_en_GB$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="AddressProvider" data-id="allowed_locales"><a href='../../charlatan/html/AddressProvider.html#method-AddressProvider-allowed_locales'><code>charlatan::AddressProvider$allowed_locales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="AddressProvider" data-id="init_person_provider"><a href='../../charlatan/html/AddressProvider.html#method-AddressProvider-init_person_provider'><code>charlatan::AddressProvider$init_person_provider()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="AddressProvider" data-id="initialize"><a href='../../charlatan/html/AddressProvider.html#method-AddressProvider-initialize'><code>charlatan::AddressProvider$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddressProvider_en_GB-address"></a>}}
\if{latex}{\out{\hypertarget{method-AddressProvider_en_GB-address}{}}}
\subsection{Method \code{address()}}{
Create an address, a combination of street, postal code and city.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddressProvider_en_GB$address()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddressProvider_en_GB-city"></a>}}
\if{latex}{\out{\hypertarget{method-AddressProvider_en_GB-city}{}}}
\subsection{Method \code{city()}}{
Create a city
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddressProvider_en_GB$city()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddressProvider_en_GB-street_name"></a>}}
\if{latex}{\out{\hypertarget{method-AddressProvider_en_GB-street_name}{}}}
\subsection{Method \code{street_name()}}{
Create a street name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddressProvider_en_GB$street_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddressProvider_en_GB-street_address"></a>}}
\if{latex}{\out{\hypertarget{method-AddressProvider_en_GB-street_address}{}}}
\subsection{Method \code{street_address()}}{
Create a street address, a combination of streetname and house indicator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddressProvider_en_GB$street_address()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddressProvider_en_GB-postcode"></a>}}
\if{latex}{\out{\hypertarget{method-AddressProvider_en_GB-postcode}{}}}
\subsection{Method \code{postcode()}}{
Create a postal code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddressProvider_en_GB$postcode()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddressProvider_en_GB-building_number"></a>}}
\if{latex}{\out{\hypertarget{method-AddressProvider_en_GB-building_number}{}}}
\subsection{Method \code{building_number()}}{
Create a building number
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddressProvider_en_GB$building_number()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AddressProvider_en_GB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AddressProvider_en_GB-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AddressProvider_en_GB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
