% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_providers.R
\docType{data}
\name{available_providers}
\alias{available_providers}
\title{Available Providers}
\format{
A vector with 22 items:
\describe{
\item{provider}{The base provider}

}
}
\description{
A vector of locales available in \pkg{charlatan}
}
\keyword{data}
