% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_card.R
\name{ch_credit}
\alias{ch_credit}
\alias{ch_credit_card_provider}
\alias{ch_credit_card_number}
\alias{ch_credit_card_security_code}
\title{Create fake credit card data}
\usage{
ch_credit_card_provider(n = 1)

ch_credit_card_number(n = 1)

ch_credit_card_security_code(n = 1)
}
\arguments{
\item{n}{(integer) number of things to get, any non-negative integer}
}
\description{
Create fake credit card data
}
\examples{
ch_credit_card_provider()
ch_credit_card_provider(n = 4)

ch_credit_card_number()
ch_credit_card_number(n = 10)
ch_credit_card_number(n = 500)

ch_credit_card_security_code()
ch_credit_card_security_code(n = 10)
ch_credit_card_security_code(n = 500)
}
\seealso{
\link{CreditCardProvider}
}
