% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{ch_currency}
\alias{ch_currencies}
\alias{ch_currency}
\title{Create fake currencies}
\usage{
ch_currency()

ch_currencies(n = 10)
}
\arguments{
\item{n}{(integer) number of things to get, 1 to inifinity}
}
\description{
Create fake currencies
}
\examples{
ch_currency()
ch_currencies(10)
ch_currencies(500)
}

