% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{ch_doi}
\alias{ch_doi}
\title{Create fake DOIs (Digital Object Identifiers)}
\usage{
ch_doi(n = 1)
}
\arguments{
\item{n}{(integer) number of things to get, any non-negative integer}
}
\description{
Create fake DOIs (Digital Object Identifiers)
}
\examples{
ch_doi()
ch_doi(10)
ch_doi(100)
}
\seealso{
\link{DOIProvider}
}
