% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{ch_generate}
\alias{ch_generate}
\title{Generate a fake dataset}
\usage{
ch_generate(..., n = 10)
}
\arguments{
\item{...}{columns to include. must be in the allowed set. See
\strong{Allowed colun names} below}

\item{n}{(integer) number of things to get, any positive integer > 1}
}
\description{
Generate a fake dataset
}
\section{Allowed column names}{

\itemize{
 \item name (default included)
 \item job (default included)
 \item phone_number (default included)
 \item currency
 \item hex_color
 \item color_name
 \item rgb_color
 \item rgb_css_color
}
}

\examples{
ch_generate()
ch_generate(n = 1)
ch_generate(n = 100)

ch_generate('job')
ch_generate('job', 'name')
ch_generate('job', 'color_name')
}
