% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isced.R
\name{ch_isced}
\alias{ch_isced}
\title{Create fake ISCED-F 2013 keywords}
\usage{
ch_isced(n = 1, locale = NULL)
}
\arguments{
\item{n}{(integer) number of things to get, any non-negative integer}

\item{locale}{(character) the locale to use. options: en_US (default),
fr_FR, nl_NL.}
}
\description{
ISCED Fields 2013 (ISCED-F 2013): International Standard Classification of Education - Fields of Education and Training
}
\examples{
ch_isced()
ch_isced(10)
ch_isced(500)

ch_isced(locale = "fr_FR", n = 10)
ch_isced(locale = "nl_NL", n = 10)
}
