% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{ch_missing}
\alias{ch_missing}
\title{Create missing data}
\usage{
ch_missing(x, n = 1)
}
\arguments{
\item{n}{(integer) number of things to get, any non-negative integer}
}
\description{
Create missing data
}
\examples{
ch_missing(letters)
ch_missing(letters, 10)
ch_missing(letters, 20)
}
