% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone_number.R
\name{ch_phone_number}
\alias{ch_phone_number}
\title{Create fake phone numbers}
\usage{
ch_phone_number(n = 1)
}
\arguments{
\item{n}{(integer) number of things to get, 1 to inifinity}
}
\description{
Create fake phone numbers
}
\examples{
ch_phone_number()
ch_phone_number(10)
ch_phone_number(500)
}
