% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{ch_sequence}
\alias{ch_sequence}
\title{Create fake gene sequences}
\usage{
ch_sequence(n = 1, length = 30)
}
\arguments{
\item{n}{(integer) number of things to get, 1 to inifinity}

\item{length}{(integer) length of sequence to create}
}
\description{
Create fake gene sequences
}
\examples{
ch_sequence()
ch_sequence(10)
ch_sequence(100)

ch_sequence(length = 500)
ch_sequence(10, length = 500)
}

