% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn.R
\name{ch_ssn}
\alias{ch_ssn}
\title{Create fake Social Security Numbers}
\usage{
ch_ssn(n = 1, locale = NULL)
}
\arguments{
\item{n}{(integer) number of things to get, any non-negative integer}

\item{locale}{(character) the locale to use. See
\code{SSNProvider$new()$allowed_locales()} for locales
supported (default: en_US)}
}
\description{
Create fake Social Security Numbers
}
\examples{
ch_ssn()
ch_ssn(10)
}
\seealso{
\link{SSNProvider}
}
