% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address.R
\name{ch_street_address}
\alias{ch_street_address}
\title{Create fake addresses}
\usage{
ch_street_address(n = 1, locale = NULL)
}
\arguments{
\item{n}{(integer) number of things to get, any non-negative integer}

\item{locale}{(character) the locale to use. See
\code{AddressProvider$new()$allowed_locales()} for locales supported.}
}
\description{
Create fake addresses
}
\examples{
ch_street_address()
ch_street_address(n = 10)
ch_postcode()
ch_postcode(n = 10)
}
\seealso{
\link{AddressProvider}
}
