% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlatan-package.R
\docType{package}
\name{charlatan-package}
\alias{charlatan}
\alias{charlatan-package}
\title{charlatan}
\description{
charlatan
}
\examples{
ch_name()
ch_phone_number()
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

