% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_locales.R
\name{charlatan_locales}
\alias{charlatan_locales}
\title{Available locales}
\description{
Vector of available locales - not all are available across all
types of data.
}
\examples{
charlatan_locales()
}

