% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checksum_util}
\alias{checksum_util}
\title{generalized util function for sequence multiplication}
\usage{
checksum_util(vector, multiplicationvector)
}
\description{
Util function for ISBN, ean, SSN providers.
for cases such as sum(x\link{4}*5 + x\link{3}*4 + x\link{2}*3 + x\link{1} *2)
}
\keyword{internal}
