% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-provider.R
\name{ColorProvider}
\alias{ColorProvider}
\title{ColorProvider}
\value{
A ColorProvider object that can generate
colors.
}
\description{
methods for colors
create color names, hex values, rgb values
or css values.
}
\seealso{
Other ParentProviders: 
\code{\link{AddressProvider}},
\code{\link{CompanyProvider}},
\code{\link{InternetProvider}},
\code{\link{JobProvider}},
\code{\link{LoremProvider}},
\code{\link{PersonProvider}},
\code{\link{PhoneNumberProvider}},
\code{\link{TaxonomyProvider}},
\code{\link{UserAgentProvider}}
}
\concept{ParentProviders}
\section{Super classes}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{ColorProvider}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{all_colors}}{(character) xxx}

\item{\code{safe_colors}}{(character) xxx}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ColorProvider-color_name}{\code{ColorProvider$color_name()}}
\item \href{#method-ColorProvider-hex_from_name}{\code{ColorProvider$hex_from_name()}}
\item \href{#method-ColorProvider-safe_color_name}{\code{ColorProvider$safe_color_name()}}
\item \href{#method-ColorProvider-hex_color}{\code{ColorProvider$hex_color()}}
\item \href{#method-ColorProvider-safe_hex_color}{\code{ColorProvider$safe_hex_color()}}
\item \href{#method-ColorProvider-rgb_color}{\code{ColorProvider$rgb_color()}}
\item \href{#method-ColorProvider-rgb_css_color}{\code{ColorProvider$rgb_css_color()}}
\item \href{#method-ColorProvider-clone}{\code{ColorProvider$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="allowed_locales"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-allowed_locales'><code>charlatan::BaseProvider$allowed_locales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="initialize"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-initialize'><code>charlatan::BaseProvider$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="print"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-print'><code>charlatan::BaseProvider$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-color_name"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-color_name}{}}}
\subsection{Method \code{color_name()}}{
color name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$color_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-hex_from_name"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-hex_from_name}{}}}
\subsection{Method \code{hex_from_name()}}{
get color by name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$hex_from_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{color name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
hex value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-safe_color_name"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-safe_color_name}{}}}
\subsection{Method \code{safe_color_name()}}{
safe color name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$safe_color_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-hex_color"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-hex_color}{}}}
\subsection{Method \code{hex_color()}}{
hex color
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$hex_color()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-safe_hex_color"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-safe_hex_color}{}}}
\subsection{Method \code{safe_hex_color()}}{
safe hex color
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$safe_hex_color()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-rgb_color"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-rgb_color}{}}}
\subsection{Method \code{rgb_color()}}{
RGB color
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$rgb_color()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-rgb_css_color"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-rgb_css_color}{}}}
\subsection{Method \code{rgb_css_color()}}{
RGB CSS color
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$rgb_css_color()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
