% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-provider-uk_UA.R
\name{ColorProvider_uk_UA}
\alias{ColorProvider_uk_UA}
\title{ColorProvider Ukrainian (Ukraine)}
\value{
A ColorProvider object that can generate
colors.
}
\description{
methods for colors
create color names, hex values, rgb values
or css values.
}
\examples{
x <- ColorProvider_uk_UA$new()
x$locale
x$color_name()
x$safe_color_name()
x$hex_color()
x$safe_hex_color()
x$rgb_color()
x$rgb_css_color()
}
\seealso{
Other uk: 
\code{\link{JobProvider_uk_UA}},
\code{\link{PhoneNumberProvider_uk_UA}},
\code{\link{ukrainian-language}}

Other UA: 
\code{\link{JobProvider_uk_UA}},
\code{\link{PhoneNumberProvider_uk_UA}}
}
\concept{UA}
\concept{uk}
\section{Super classes}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{\link[charlatan:ColorProvider]{charlatan::ColorProvider}} -> \code{ColorProvider_uk_UA}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ColorProvider_uk_UA-clone}{\code{ColorProvider_uk_UA$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="allowed_locales"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-allowed_locales'><code>charlatan::BaseProvider$allowed_locales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="initialize"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-initialize'><code>charlatan::BaseProvider$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="print"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-print'><code>charlatan::BaseProvider$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="ColorProvider" data-id="color_name"><a href='../../charlatan/html/ColorProvider.html#method-ColorProvider-color_name'><code>charlatan::ColorProvider$color_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="ColorProvider" data-id="hex_color"><a href='../../charlatan/html/ColorProvider.html#method-ColorProvider-hex_color'><code>charlatan::ColorProvider$hex_color()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="ColorProvider" data-id="hex_from_name"><a href='../../charlatan/html/ColorProvider.html#method-ColorProvider-hex_from_name'><code>charlatan::ColorProvider$hex_from_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="ColorProvider" data-id="rgb_color"><a href='../../charlatan/html/ColorProvider.html#method-ColorProvider-rgb_color'><code>charlatan::ColorProvider$rgb_color()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="ColorProvider" data-id="rgb_css_color"><a href='../../charlatan/html/ColorProvider.html#method-ColorProvider-rgb_css_color'><code>charlatan::ColorProvider$rgb_css_color()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="ColorProvider" data-id="safe_color_name"><a href='../../charlatan/html/ColorProvider.html#method-ColorProvider-safe_color_name'><code>charlatan::ColorProvider$safe_color_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="ColorProvider" data-id="safe_hex_color"><a href='../../charlatan/html/ColorProvider.html#method-ColorProvider-safe_hex_color'><code>charlatan::ColorProvider$safe_hex_color()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider_uk_UA-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider_uk_UA-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider_uk_UA$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
