% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company-provider.R
\name{company_provider}
\alias{company_provider}
\title{CompanyProvider}
\arguments{
\item{locale}{(character) the locale to use. options: en_US (default),
fr_FR.}
}
\description{
CompanyProvider
}
\details{
\strong{Methods}
  \describe{
   \item{\code{company}}{
     a company name
   }
   \item{\code{company_suffix}}{
     a company suffix
   }
 }
}
\examples{
x <- company_provider('en_US')
x$locale
x$company()
x$company_suffix()
x$catch_phrase()
x$bs()

x <- company_provider(locale = "fr_FR")
x$locale
x$company()
x$company_suffix()
x$siren()

x <- company_provider(locale = "hr_HR")
x$locale
x$company()
x$company_suffix()

x <- company_provider(locale = "it_IT")
x$locale
x$company()
x$company_suffix()
x$bs()

company_provider(locale = "es_MX")$bs()
company_provider(locale = "es_MX")$company_prefix()
company_provider(locale = "es_MX")$catch_phrase()

company_provider(locale = "bg_BG")$company()
company_provider(locale = "cs_CZ")$company()
company_provider(locale = "de_DE")$company()
company_provider(locale = "fa_IR")$company()
}
\keyword{internal}

