% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company-provider-fa_IR.R
\name{CompanyProvider_fa_IR}
\alias{CompanyProvider_fa_IR}
\title{CompanyProvider fa_IR}
\description{
company methods for locale Persian (Iran) (fa_IR).
}
\examples{
x <- company("fa_IR")
x$locale
x$company()
x$catch_phrase()
x$bs()
}
\seealso{
Other fa: 
\code{\link{InternetProvider_fa_IR}},
\code{\link{JobProvider_fa_IR}},
\code{\link{PersonProvider_fa_IR}},
\code{\link{PhoneNumberProvider_fa_IR}},
\code{\link{farsi-language}}

Other IR: 
\code{\link{InternetProvider_fa_IR}},
\code{\link{JobProvider_fa_IR}},
\code{\link{PersonProvider_fa_IR}},
\code{\link{PhoneNumberProvider_fa_IR}}
}
\concept{IR}
\concept{fa}
\section{Super classes}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{\link[charlatan:CompanyProvider]{charlatan::CompanyProvider}} -> \code{CompanyProvider_fa_IR}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{locale}}{(character) xxx}

\item{\code{company_names}}{vector of company names.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CompanyProvider_fa_IR-company}{\code{CompanyProvider_fa_IR$company()}}
\item \href{#method-CompanyProvider_fa_IR-clone}{\code{CompanyProvider_fa_IR$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="CompanyProvider" data-id="allowed_locales"><a href='../../charlatan/html/CompanyProvider.html#method-CompanyProvider-allowed_locales'><code>charlatan::CompanyProvider$allowed_locales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="CompanyProvider" data-id="bs"><a href='../../charlatan/html/CompanyProvider.html#method-CompanyProvider-bs'><code>charlatan::CompanyProvider$bs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="CompanyProvider" data-id="catch_phrase"><a href='../../charlatan/html/CompanyProvider.html#method-CompanyProvider-catch_phrase'><code>charlatan::CompanyProvider$catch_phrase()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="CompanyProvider" data-id="initialize"><a href='../../charlatan/html/CompanyProvider.html#method-CompanyProvider-initialize'><code>charlatan::CompanyProvider$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="CompanyProvider" data-id="siren"><a href='../../charlatan/html/CompanyProvider.html#method-CompanyProvider-siren'><code>charlatan::CompanyProvider$siren()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanyProvider_fa_IR-company"></a>}}
\if{latex}{\out{\hypertarget{method-CompanyProvider_fa_IR-company}{}}}
\subsection{Method \code{company()}}{
a company name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanyProvider_fa_IR$company()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanyProvider_fa_IR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CompanyProvider_fa_IR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanyProvider_fa_IR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
