% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company-provider-fr_FR.R
\name{CompanyProvider_fr_FR}
\alias{CompanyProvider_fr_FR}
\title{CompanyProvider fr_FR}
\description{
company methods for locale French France(fr_FR).
}
\examples{
x <- CompanyProvider_fr_FR$new()
x$locale
x$company()
x$catch_phrase()
}
\seealso{
Other FR: 
\code{\link{InternetProvider_fr_FR}},
\code{\link{JobProvider_fr_FR}},
\code{\link{PersonProvider_fr_FR}},
\code{\link{PhoneNumberProvider_fr_FR}}

Other fr: 
\code{\link{InternetProvider_fr_FR}},
\code{\link{JobProvider_fr_CH}},
\code{\link{JobProvider_fr_FR}},
\code{\link{PersonProvider_fr_CH}},
\code{\link{PersonProvider_fr_FR}},
\code{\link{PhoneNumberProvider_fr_CH}},
\code{\link{PhoneNumberProvider_fr_FR}},
\code{\link{french-language}}
}
\concept{FR}
\concept{fr}
\section{Super classes}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{\link[charlatan:CompanyProvider]{charlatan::CompanyProvider}} -> \code{CompanyProvider_fr_FR}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catch_phrase_formats}}{formatting for catch phrases.}

\item{\code{company_formats}}{formatting for company names}

\item{\code{nouns}}{nouns}

\item{\code{verbs}}{verbs}

\item{\code{attributes}}{attributes}

\item{\code{company_suffixes}}{endings of companies.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CompanyProvider_fr_FR-catch_phrase}{\code{CompanyProvider_fr_FR$catch_phrase()}}
\item \href{#method-CompanyProvider_fr_FR-clone}{\code{CompanyProvider_fr_FR$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="print"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-print'><code>charlatan::BaseProvider$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="CompanyProvider" data-id="allowed_locales"><a href='../../charlatan/html/CompanyProvider.html#method-CompanyProvider-allowed_locales'><code>charlatan::CompanyProvider$allowed_locales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="CompanyProvider" data-id="bs"><a href='../../charlatan/html/CompanyProvider.html#method-CompanyProvider-bs'><code>charlatan::CompanyProvider$bs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="CompanyProvider" data-id="company"><a href='../../charlatan/html/CompanyProvider.html#method-CompanyProvider-company'><code>charlatan::CompanyProvider$company()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="CompanyProvider" data-id="initialize"><a href='../../charlatan/html/CompanyProvider.html#method-CompanyProvider-initialize'><code>charlatan::CompanyProvider$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanyProvider_fr_FR-catch_phrase"></a>}}
\if{latex}{\out{\hypertarget{method-CompanyProvider_fr_FR-catch_phrase}{}}}
\subsection{Method \code{catch_phrase()}}{
generate a catch phrase for a company.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanyProvider_fr_FR$catch_phrase()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CompanyProvider_fr_FR-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CompanyProvider_fr_FR-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CompanyProvider_fr_FR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
