% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\alias{ch_lon}
\alias{ch_lat}
\alias{ch_position}
\title{Create fake coordinates}
\usage{
ch_lon(n = 1)

ch_lat(n = 1)

ch_position(n = 1, bbox = NULL)
}
\arguments{
\item{n}{(integer) number of things to get, 1 to inifinity}

\item{bbox}{a bounding box of the form [w,s,e,n]}
}
\description{
Create fake coordinates
}
\examples{
ch_lon()
ch_lon(10)

ch_lat()
ch_lat(10)

ch_position()
ch_position(10)
}
