% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_card-provider.R
\docType{data}
\name{CreditCardProvider}
\alias{CreditCardProvider}
\title{CreditCardProvider}
\description{
CreditCardProvider
}
\details{
\strong{Methods}
\itemize{
\item \code{credit_card_type(card_type)} - credit card type
\item \code{credit_card_provider(prefix, length)} - credit card provider
\item \code{credit_card_number(card_type)} - credit card number
\item \code{credit_card_security_code(card_type)} - credit card security code
}
}
\examples{
z <- CreditCardProvider$new()
z$credit_card_provider()
z$credit_card_number()
z$credit_card_security_code()
}
\keyword{internal}
