% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time.R
\name{date_time}
\alias{date_time}
\alias{ch_timezone}
\alias{ch_unix_time}
\alias{ch_date_time}
\title{Create dates and times}
\usage{
ch_timezone(n = 1)

ch_unix_time(n = 1)

ch_date_time(n = 1)
}
\arguments{
\item{n}{(integer) number of things to get, 1 to inifinity}
}
\description{
Create dates and times
}
\examples{
ch_timezone()
ch_timezone(10)

ch_unix_time()
ch_unix_time(20)

ch_date_time()
ch_date_time(20)
}
