% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{dutch-language}
\alias{dutch-language}
\title{Dutch Language}
\description{
Providers with the Dutch locale (nl).
}
\seealso{
Other nl: 
\code{\link{AddressProvider_nl_NL}},
\code{\link{ElementProvider_nl_NL}},
\code{\link{JobProvider_nl_NL}},
\code{\link{PersonProvider_nl_NL}},
\code{\link{PhoneNumberProvider_nl_BE}},
\code{\link{PhoneNumberProvider_nl_NL}},
\code{\link{SSNProvider_nl_NL}}

Other languages: 
\code{\link{arabic-language}},
\code{\link{bosnian-language}},
\code{\link{bulgarian-language}},
\code{\link{chinese-language}},
\code{\link{croatian-language}},
\code{\link{czech-language}},
\code{\link{danish-language}},
\code{\link{english-language}},
\code{\link{farsi-language}},
\code{\link{finnish-language}},
\code{\link{french-language}},
\code{\link{german-language}},
\code{\link{greek-language}},
\code{\link{hebrew-language}},
\code{\link{hindi-language}},
\code{\link{hungarian-language}},
\code{\link{indonesian-language}},
\code{\link{italian-language}},
\code{\link{japanese-language}},
\code{\link{korean-language}},
\code{\link{latin-language}},
\code{\link{latvian-language}},
\code{\link{lithuanian-language}},
\code{\link{norwegian-language}},
\code{\link{polish-language}},
\code{\link{portuguese-language}},
\code{\link{russian-language}},
\code{\link{spanish-language}},
\code{\link{swedish-language}},
\code{\link{thai-language}},
\code{\link{turkish-language}},
\code{\link{ukrainian-language}}
}
\concept{languages}
\concept{nl}
