% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-provider.R
\name{ElementProvider_nl_NL}
\alias{ElementProvider_nl_NL}
\title{ElementProvider for the Netherlands}
\description{
chemical elements methods
}
\details{
Data from Wikipedia at
\url{https://nl.wikipedia.org/wiki/Lijst_van_chemische_elementen}
}
\examples{
z <- ElementProvider_nl_NL$new()
z$symbol()
z$element()
}
\seealso{
Other nl: 
\code{\link{AddressProvider_nl_NL}},
\code{\link{JobProvider_nl_NL}},
\code{\link{PersonProvider_nl_NL}},
\code{\link{PhoneNumberProvider_nl_BE}},
\code{\link{PhoneNumberProvider_nl_NL}},
\code{\link{SSNProvider_nl_NL}},
\code{\link{dutch-language}}

Other NL: 
\code{\link{AddressProvider_nl_NL}},
\code{\link{JobProvider_nl_NL}},
\code{\link{PersonProvider_nl_NL}},
\code{\link{PhoneNumberProvider_nl_NL}},
\code{\link{SSNProvider_nl_NL}}
}
\concept{NL}
\concept{nl}
\section{Super classes}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{\link[charlatan:ElementProvider]{charlatan::ElementProvider}} -> \code{ElementProvider_nl_NL}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ElementProvider_nl_NL-clone}{\code{ElementProvider_nl_NL$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="allowed_locales"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-allowed_locales'><code>charlatan::BaseProvider$allowed_locales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="initialize"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-initialize'><code>charlatan::BaseProvider$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="print"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-print'><code>charlatan::BaseProvider$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="ElementProvider" data-id="element"><a href='../../charlatan/html/ElementProvider.html#method-ElementProvider-element'><code>charlatan::ElementProvider$element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="ElementProvider" data-id="element_by_number"><a href='../../charlatan/html/ElementProvider.html#method-ElementProvider-element_by_number'><code>charlatan::ElementProvider$element_by_number()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="ElementProvider" data-id="elements"><a href='../../charlatan/html/ElementProvider.html#method-ElementProvider-elements'><code>charlatan::ElementProvider$elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="ElementProvider" data-id="symbol"><a href='../../charlatan/html/ElementProvider.html#method-ElementProvider-symbol'><code>charlatan::ElementProvider$symbol()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="ElementProvider" data-id="symbol_by_number"><a href='../../charlatan/html/ElementProvider.html#method-ElementProvider-symbol_by_number'><code>charlatan::ElementProvider$symbol_by_number()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementProvider_nl_NL-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ElementProvider_nl_NL-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementProvider_nl_NL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
