% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{elements}
\alias{elements}
\alias{ch_element_symbol}
\alias{ch_element_element}
\title{Get elements}
\usage{
ch_element_symbol(n = 1)

ch_element_element(n = 1)
}
\arguments{
\item{n}{(integer) number of things to get, any non-negative integer}
}
\description{
Get elements
}
\examples{
ch_element_symbol()
ch_element_symbol(10)
ch_element_symbol(50)

ch_element_element()
ch_element_element(10)
ch_element_element(50)
}
