% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{english-language}
\alias{english-language}
\title{English Language}
\description{
Providers with the English locale (en).
}
\seealso{
Other en: 
\code{\link{AddressProvider_en_GB}},
\code{\link{AddressProvider_en_NZ}},
\code{\link{ColorProvider_en_US}},
\code{\link{CompanyProvider_en_US}},
\code{\link{ElementProvider_en_US}},
\code{\link{InternetProvider_en_AU}},
\code{\link{InternetProvider_en_US}},
\code{\link{JobProvider_en_US}},
\code{\link{LoremProvider_en_US}},
\code{\link{PersonProvider_en_GB}},
\code{\link{PersonProvider_en_NZ}},
\code{\link{PersonProvider_en_US}},
\code{\link{PhoneNumberProvider_en_AU}},
\code{\link{PhoneNumberProvider_en_CA}},
\code{\link{PhoneNumberProvider_en_GB}},
\code{\link{PhoneNumberProvider_en_NZ}},
\code{\link{PhoneNumberProvider_en_US}},
\code{\link{PhoneNumberProvider_es_MX}},
\code{\link{PhoneNumberProvider_tr_TR}},
\code{\link{PhoneNumberProvider}},
\code{\link{SSNProvider_en_US}},
\code{\link{TaxonomyProvider_en_US}},
\code{\link{UserAgentProvider_en_US}},
\code{\link{UserAgentProvider}}

Other languages: 
\code{\link{arabic-language}},
\code{\link{bosnian-language}},
\code{\link{bulgarian-language}},
\code{\link{chinese-language}},
\code{\link{croatian-language}},
\code{\link{czech-language}},
\code{\link{danish-language}},
\code{\link{dutch-language}},
\code{\link{farsi-language}},
\code{\link{finnish-language}},
\code{\link{french-language}},
\code{\link{german-language}},
\code{\link{greek-language}},
\code{\link{hebrew-language}},
\code{\link{hindi-language}},
\code{\link{hungarian-language}},
\code{\link{indonesian-language}},
\code{\link{italian-language}},
\code{\link{japanese-language}},
\code{\link{korean-language}},
\code{\link{latin-language}},
\code{\link{latvian-language}},
\code{\link{lithuanian-language}},
\code{\link{norwegian-language}},
\code{\link{polish-language}},
\code{\link{portuguese-language}},
\code{\link{russian-language}},
\code{\link{spanish-language}},
\code{\link{swedish-language}},
\code{\link{thai-language}},
\code{\link{turkish-language}},
\code{\link{ukrainian-language}}
}
\concept{en}
\concept{languages}
