% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-provider.R
\docType{data}
\name{FileProvider}
\alias{FileProvider}
\alias{file_provider_locales}
\title{FileProvider}
\format{An object of class \code{character} of length 1.}
\usage{
file_provider_locales
}
\description{
FileProvider
}
\details{
\strong{Methods}
  \describe{
   \item{\code{to_ascii}}{
    convert to ascii
  }
 }
}
\examples{
(x <- FileProvider$new())
x$locale 
x$mime_type()
x$file_extension()
}
\keyword{datasets}
\keyword{internal}
