% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraudster.R
\name{fraudster}
\alias{fraudster}
\title{Fraudster - catch all client to make all types of fake data}
\usage{
fraudster(locale = NULL)
}
\arguments{
\item{locale}{(character) the locale to use. options: en_US (default),
fr_FR, fr_CH, hr_FR, fa_IR, pl_PL, ru_RU, uk_UA, zh_TW.}
}
\description{
Fraudster - catch all client to make all types of fake data
}
\examples{
# English - the default locale
(x <- fraudster())
x$job()
x$name()
x$color_name()
x$safe_color_name()
x$hex_color()
x$safe_hex_color()
x$rgb_color()
x$rgb_css_color()

# different locales
## French
(y <- fraudster(locale = "fr_FR"))
y$job()

## Croatian
(z <- fraudster(locale = "hr_HR"))
z$job()

## Ukranian
(w <- fraudster(locale = "uk_UA"))
w$job()
w$color_name()
}

