% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraudster.R
\name{FraudsterClient}
\alias{FraudsterClient}
\title{FraudsterClient}
\description{
Fraudster R6 client
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{locale}}{(character) the locale to use}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FraudsterClient-new}{\code{FraudsterClient$new()}}
\item \href{#method-FraudsterClient-print}{\code{FraudsterClient$print()}}
\item \href{#method-FraudsterClient-job}{\code{FraudsterClient$job()}}
\item \href{#method-FraudsterClient-name}{\code{FraudsterClient$name()}}
\item \href{#method-FraudsterClient-color_name}{\code{FraudsterClient$color_name()}}
\item \href{#method-FraudsterClient-safe_color_name}{\code{FraudsterClient$safe_color_name()}}
\item \href{#method-FraudsterClient-hex_color}{\code{FraudsterClient$hex_color()}}
\item \href{#method-FraudsterClient-safe_hex_color}{\code{FraudsterClient$safe_hex_color()}}
\item \href{#method-FraudsterClient-rgb_color}{\code{FraudsterClient$rgb_color()}}
\item \href{#method-FraudsterClient-rgb_css_color}{\code{FraudsterClient$rgb_css_color()}}
\item \href{#method-FraudsterClient-lat}{\code{FraudsterClient$lat()}}
\item \href{#method-FraudsterClient-lon}{\code{FraudsterClient$lon()}}
\item \href{#method-FraudsterClient-position}{\code{FraudsterClient$position()}}
\item \href{#method-FraudsterClient-doi}{\code{FraudsterClient$doi()}}
\item \href{#method-FraudsterClient-timezone}{\code{FraudsterClient$timezone()}}
\item \href{#method-FraudsterClient-unix_time}{\code{FraudsterClient$unix_time()}}
\item \href{#method-FraudsterClient-date_time}{\code{FraudsterClient$date_time()}}
\item \href{#method-FraudsterClient-genus}{\code{FraudsterClient$genus()}}
\item \href{#method-FraudsterClient-epithet}{\code{FraudsterClient$epithet()}}
\item \href{#method-FraudsterClient-species}{\code{FraudsterClient$species()}}
\item \href{#method-FraudsterClient-sequence}{\code{FraudsterClient$sequence()}}
\item \href{#method-FraudsterClient-phone_number}{\code{FraudsterClient$phone_number()}}
\item \href{#method-FraudsterClient-double}{\code{FraudsterClient$double()}}
\item \href{#method-FraudsterClient-integer}{\code{FraudsterClient$integer()}}
\item \href{#method-FraudsterClient-uniform}{\code{FraudsterClient$uniform()}}
\item \href{#method-FraudsterClient-norm}{\code{FraudsterClient$norm()}}
\item \href{#method-FraudsterClient-lnorm}{\code{FraudsterClient$lnorm()}}
\item \href{#method-FraudsterClient-beta}{\code{FraudsterClient$beta()}}
\item \href{#method-FraudsterClient-currency}{\code{FraudsterClient$currency()}}
\item \href{#method-FraudsterClient-credit_card_provider}{\code{FraudsterClient$credit_card_provider()}}
\item \href{#method-FraudsterClient-credit_card_number}{\code{FraudsterClient$credit_card_number()}}
\item \href{#method-FraudsterClient-credit_card_security_code}{\code{FraudsterClient$credit_card_security_code()}}
\item \href{#method-FraudsterClient-clone}{\code{FraudsterClient$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{FraudsterClient} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$new(locale = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{locale}}{(character) the locale to use. options: en_US (default),
fr_FR, fr_CH, hr_FR, fa_IR, pl_PL, ru_RU, uk_UA, zh_TW.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{RequestSignature} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-print"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-print}{}}}
\subsection{Method \code{print()}}{
print method for the \code{FraudsterClient} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$print(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{self}

\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-job"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-job}{}}}
\subsection{Method \code{job()}}{
jobs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$job(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-name"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-name}{}}}
\subsection{Method \code{name()}}{
names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$name(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-color_name"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-color_name}{}}}
\subsection{Method \code{color_name()}}{
colors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$color_name(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-safe_color_name"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-safe_color_name}{}}}
\subsection{Method \code{safe_color_name()}}{
safe color name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$safe_color_name(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-hex_color"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-hex_color}{}}}
\subsection{Method \code{hex_color()}}{
hex color
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$hex_color(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-safe_hex_color"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-safe_hex_color}{}}}
\subsection{Method \code{safe_hex_color()}}{
safe hex color
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$safe_hex_color(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-rgb_color"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-rgb_color}{}}}
\subsection{Method \code{rgb_color()}}{
rgb color
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$rgb_color(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-rgb_css_color"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-rgb_css_color}{}}}
\subsection{Method \code{rgb_css_color()}}{
rgb css color
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$rgb_css_color(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-lat"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-lat}{}}}
\subsection{Method \code{lat()}}{
latitude
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$lat(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-lon"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-lon}{}}}
\subsection{Method \code{lon()}}{
longitude
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$lon(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-position"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-position}{}}}
\subsection{Method \code{position()}}{
long/lat coordinate pair
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$position(n = 1, bbox = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}

\item{\code{bbox}}{a bounding box, see \code{\link[=ch_position]{ch_position()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-doi"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-doi}{}}}
\subsection{Method \code{doi()}}{
DOIs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$doi(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-timezone"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-timezone}{}}}
\subsection{Method \code{timezone()}}{
date times
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$timezone(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-unix_time"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-unix_time}{}}}
\subsection{Method \code{unix_time()}}{
unix time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$unix_time(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-date_time"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-date_time}{}}}
\subsection{Method \code{date_time()}}{
date time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$date_time(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-genus"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-genus}{}}}
\subsection{Method \code{genus()}}{
taxonomic genus
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$genus(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-epithet"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-epithet}{}}}
\subsection{Method \code{epithet()}}{
taxonomic epithet
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$epithet(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-species"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-species}{}}}
\subsection{Method \code{species()}}{
taxonomic species (genus + epithet)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$species(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-sequence"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-sequence}{}}}
\subsection{Method \code{sequence()}}{
random genetic sequence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$sequence(n = 1, length = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}

\item{\code{length}}{(integer) length of the sequence. default: 30}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-phone_number"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-phone_number}{}}}
\subsection{Method \code{phone_number()}}{
phone number
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$phone_number(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-double"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-double}{}}}
\subsection{Method \code{double()}}{
a double
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$double(n = 1, mean = 0, sd = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}

\item{\code{mean}}{mean value, default: 0}

\item{\code{sd}}{standard deviation, default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-integer"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-integer}{}}}
\subsection{Method \code{integer()}}{
an integer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$integer(n = 1, min = 1, max = 1000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}

\item{\code{min}}{minimum value, default: 1}

\item{\code{max}}{maximum value, default: 1000}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-uniform"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-uniform}{}}}
\subsection{Method \code{uniform()}}{
an integer from a uniform distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$uniform(n = 1, min = 0, max = 9999)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}

\item{\code{min}}{minimum value, default: 0}

\item{\code{max}}{maximum value, default: 9999}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-norm"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-norm}{}}}
\subsection{Method \code{norm()}}{
an integer from a normal distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$norm(n = 1, mean = 0, sd = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}

\item{\code{mean}}{mean value, default: 0}

\item{\code{sd}}{standard deviation, default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-lnorm"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-lnorm}{}}}
\subsection{Method \code{lnorm()}}{
an integer from a lognormal distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$lnorm(n = 1, mean = 0, sd = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}

\item{\code{mean}}{mean value, default: 0}

\item{\code{sd}}{standard deviation, default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-beta"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-beta}{}}}
\subsection{Method \code{beta()}}{
an integer from a beta distribution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$beta(n = 1, shape1, shape2, ncp = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}

\item{\code{shape1}}{non-negative parameters of the Beta distribution}

\item{\code{shape2}}{non-negative parameters of the Beta distribution}

\item{\code{ncp}}{non-centrality parameter, default: 0}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-currency"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-currency}{}}}
\subsection{Method \code{currency()}}{
currency
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$currency(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-credit_card_provider"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-credit_card_provider}{}}}
\subsection{Method \code{credit_card_provider()}}{
credit card provider
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$credit_card_provider(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-credit_card_number"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-credit_card_number}{}}}
\subsection{Method \code{credit_card_number()}}{
credit card number
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$credit_card_number(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-credit_card_security_code"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-credit_card_security_code}{}}}
\subsection{Method \code{credit_card_security_code()}}{
credit card security code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$credit_card_security_code(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of random things to generaate. an integer; default: 1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FraudsterClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FraudsterClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FraudsterClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
