% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{french-language}
\alias{french-language}
\title{French Language}
\description{
Providers with the French locale (fr).
}
\seealso{
Other fr: 
\code{\link{CompanyProvider_fr_FR}},
\code{\link{InternetProvider_fr_FR}},
\code{\link{JobProvider_fr_CH}},
\code{\link{JobProvider_fr_FR}},
\code{\link{PersonProvider_fr_CH}},
\code{\link{PersonProvider_fr_FR}},
\code{\link{PhoneNumberProvider_fr_CH}},
\code{\link{PhoneNumberProvider_fr_FR}}

Other languages: 
\code{\link{arabic-language}},
\code{\link{bosnian-language}},
\code{\link{bulgarian-language}},
\code{\link{chinese-language}},
\code{\link{croatian-language}},
\code{\link{czech-language}},
\code{\link{danish-language}},
\code{\link{dutch-language}},
\code{\link{english-language}},
\code{\link{farsi-language}},
\code{\link{finnish-language}},
\code{\link{german-language}},
\code{\link{greek-language}},
\code{\link{hebrew-language}},
\code{\link{hindi-language}},
\code{\link{hungarian-language}},
\code{\link{indonesian-language}},
\code{\link{italian-language}},
\code{\link{japanese-language}},
\code{\link{korean-language}},
\code{\link{latin-language}},
\code{\link{latvian-language}},
\code{\link{lithuanian-language}},
\code{\link{norwegian-language}},
\code{\link{polish-language}},
\code{\link{portuguese-language}},
\code{\link{russian-language}},
\code{\link{spanish-language}},
\code{\link{swedish-language}},
\code{\link{thai-language}},
\code{\link{turkish-language}},
\code{\link{ukrainian-language}}
}
\concept{fr}
\concept{languages}
