% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internet-provider-en_NZ.R
\name{InternetProvider_en_NZ}
\alias{InternetProvider_en_NZ}
\title{Internet provider for New-Zealand}
\description{
methods for internet related data, like email addresses,
usernames, and websites.
}
\examples{
x <- InternetProvider_en_NZ$new()
x$email()
x$free_email()
x$mac_address()
x$company_email()
}
\seealso{
Other bg: 
\code{\link{CompanyProvider_bg_BG}},
\code{\link{InternetProvider_bg_BG}},
\code{\link{PersonProvider_bg_BG}},
\code{\link{PhoneNumberProvider_bg_BG}},
\code{\link{bulgarian-language}}

Other BG: 
\code{\link{CompanyProvider_bg_BG}},
\code{\link{InternetProvider_bg_BG}},
\code{\link{PersonProvider_bg_BG}},
\code{\link{PhoneNumberProvider_bg_BG}}
}
\concept{BG}
\concept{bg}
\section{Super classes}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{\link[charlatan:InternetProvider]{charlatan::InternetProvider}} -> \code{InternetProvider_en_NZ}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{locale}}{(character) the locale}

\item{\code{free_email_domains}}{(character) free email tlds}

\item{\code{tlds}}{(character) tlds}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InternetProvider_en_NZ-clone}{\code{InternetProvider_en_NZ$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="allowed_locales"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-allowed_locales'><code>charlatan::InternetProvider$allowed_locales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="ascii_company_email"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-ascii_company_email'><code>charlatan::InternetProvider$ascii_company_email()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="ascii_email"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-ascii_email'><code>charlatan::InternetProvider$ascii_email()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="ascii_free_email"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-ascii_free_email'><code>charlatan::InternetProvider$ascii_free_email()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="ascii_safe_email"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-ascii_safe_email'><code>charlatan::InternetProvider$ascii_safe_email()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="company_email"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-company_email'><code>charlatan::InternetProvider$company_email()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="domain_name"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-domain_name'><code>charlatan::InternetProvider$domain_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="domain_word"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-domain_word'><code>charlatan::InternetProvider$domain_word()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="email"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-email'><code>charlatan::InternetProvider$email()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="free_email"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-free_email'><code>charlatan::InternetProvider$free_email()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="free_email_domain"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-free_email_domain'><code>charlatan::InternetProvider$free_email_domain()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="image_url"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-image_url'><code>charlatan::InternetProvider$image_url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="initialize"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-initialize'><code>charlatan::InternetProvider$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="ipv4"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-ipv4'><code>charlatan::InternetProvider$ipv4()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="ipv6"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-ipv6'><code>charlatan::InternetProvider$ipv6()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="mac_address"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-mac_address'><code>charlatan::InternetProvider$mac_address()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="safe_email"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-safe_email'><code>charlatan::InternetProvider$safe_email()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="slug"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-slug'><code>charlatan::InternetProvider$slug()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="tld"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-tld'><code>charlatan::InternetProvider$tld()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="to_ascii"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-to_ascii'><code>charlatan::InternetProvider$to_ascii()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="uri"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-uri'><code>charlatan::InternetProvider$uri()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="uri_extension"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-uri_extension'><code>charlatan::InternetProvider$uri_extension()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="uri_page"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-uri_page'><code>charlatan::InternetProvider$uri_page()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="uri_path"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-uri_path'><code>charlatan::InternetProvider$uri_path()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="url"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-url'><code>charlatan::InternetProvider$url()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="InternetProvider" data-id="user_name"><a href='../../charlatan/html/InternetProvider.html#method-InternetProvider-user_name'><code>charlatan::InternetProvider$user_name()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InternetProvider_en_NZ-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InternetProvider_en_NZ-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider_en_NZ$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
