% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isced-provider.R
\docType{data}
\name{ISCEDProvider}
\alias{ISCEDProvider}
\alias{ISCED_provider_locales}
\title{ISCEDProvider}
\format{An object of class \code{character} of length 3.}
\usage{
ISCED_provider_locales
}
\arguments{
\item{locale}{(character) the locale to use. options: en_US (default),
fr_FR, nl_NL.}
}
\description{
ISCEDProvider
}
\details{
\strong{Methods}
  \describe{
    \item{\code{render()}}{
      Make a ISCED keyword
    }
  }
}
\examples{
z <- ISCEDProvider$new()
z$render()

z <- ISCEDProvider$new(locale = "fr_FR")
z$locale
z$render()

z <- ISCEDProvider$new(locale = "nl_NL")
z$locale
z$render()
}
\keyword{datasets}
\keyword{internal}
