% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{locale_mismatch}
\alias{locale_mismatch}
\title{Find mismatch between parent and child locales}
\usage{
locale_mismatch(parent_provider, child_provider)
}
\arguments{
\item{parent_provider}{provider that you want the locales to check for}

\item{child_provider}{provider where we check if parent locales are in.
For example InternetProvider has en_AU but LoremProvider does not.}
}
\value{
locales that ARE in the parent, but are not available in the child.
}
\description{
Find mismatch between parent and child locales
}
\keyword{internal}
