% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorem-provider-zh_CN.R
\name{LoremProvider_zh_CN}
\alias{LoremProvider_zh_CN}
\title{Lorem provider}
\description{
Methods for Lorem Ipsum generation.
Lorem Ipsum is a placeholder text commonly used to demonstrate the visual
form of a document or a typeface without relying on meaningful content.
}
\examples{
x <- LoremProvider_zh_CN$new()
x$word()
x$words(3)
x$words(6)
x$sentence()
x$paragraph()
x$paragraphs(3)
cat(x$paragraphs(6), sep = "\n")
x$text(19)
x <- LoremProvider_zh_CN$new(word_connector = " --- ")
x$paragraph(4)
}
\seealso{
Other zh: 
\code{\link{JobProvider_zh_TW}},
\code{\link{LoremProvider_zh_TW}},
\code{\link{PhoneNumberProvider_zh_TW}},
\code{\link{chinese-language}}
}
\concept{CN}
\concept{zh}
\section{Super classes}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{\link[charlatan:LoremProvider]{charlatan::LoremProvider}} -> \code{LoremProvider_zh_CN}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{locale}}{(character) the locale}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LoremProvider_zh_CN-clone}{\code{LoremProvider_zh_CN$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="LoremProvider" data-id="allowed_locales"><a href='../../charlatan/html/LoremProvider.html#method-LoremProvider-allowed_locales'><code>charlatan::LoremProvider$allowed_locales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="LoremProvider" data-id="initialize"><a href='../../charlatan/html/LoremProvider.html#method-LoremProvider-initialize'><code>charlatan::LoremProvider$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="LoremProvider" data-id="paragraph"><a href='../../charlatan/html/LoremProvider.html#method-LoremProvider-paragraph'><code>charlatan::LoremProvider$paragraph()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="LoremProvider" data-id="paragraphs"><a href='../../charlatan/html/LoremProvider.html#method-LoremProvider-paragraphs'><code>charlatan::LoremProvider$paragraphs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="LoremProvider" data-id="sentence"><a href='../../charlatan/html/LoremProvider.html#method-LoremProvider-sentence'><code>charlatan::LoremProvider$sentence()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="LoremProvider" data-id="sentences"><a href='../../charlatan/html/LoremProvider.html#method-LoremProvider-sentences'><code>charlatan::LoremProvider$sentences()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="LoremProvider" data-id="text"><a href='../../charlatan/html/LoremProvider.html#method-LoremProvider-text'><code>charlatan::LoremProvider$text()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="LoremProvider" data-id="word"><a href='../../charlatan/html/LoremProvider.html#method-LoremProvider-word'><code>charlatan::LoremProvider$word()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="LoremProvider" data-id="words"><a href='../../charlatan/html/LoremProvider.html#method-LoremProvider-words'><code>charlatan::LoremProvider$words()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoremProvider_zh_CN-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LoremProvider_zh_CN-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoremProvider_zh_CN$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
