% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing-data-provider.R
\docType{data}
\name{MissingDataProvider}
\alias{MissingDataProvider}
\title{MissingDataProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
MissingDataProvider
}
\description{
MissingDataProvider
}
\examples{
z <- MissingDataProvider$new()
z$make_missing(letters)
}
\keyword{internal}

