% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics.R
\name{numerics}
\alias{ch_beta}
\alias{ch_double}
\alias{ch_integer}
\alias{ch_lnorm}
\alias{ch_norm}
\alias{ch_unif}
\alias{numerics}
\title{Create numbers}
\usage{
ch_double(n = 1, mean = 0, sd = 1)

ch_double(n = 1, mean = 0, sd = 1)

ch_integer(n = 1, min = 1, max = 1000)

ch_unif(n = 1, min = 0, max = 9999)

ch_norm(n = 1, mean = 0, sd = 1)

ch_lnorm(n = 1, mean = 0, sd = 1)

ch_beta(n = 1, shape1, shape2, ncp = 0)
}
\arguments{
\item{n}{(integer) number of things to get, 1 to inifinity}

\item{mean}{mean value}

\item{sd}{standard deviation}

\item{min}{minimum value}

\item{max}{maximum value}

\item{shape1, shape2}{non-negative parameters of the Beta distribution}

\item{ncp}{non-centrality parameter}
}
\description{
Create numbers
}
\examples{
ch_double()
ch_double(10)
ch_double(100)

ch_integer()
ch_integer(10)
ch_integer(100)

ch_unif()
ch_norm()
ch_lnorm()
ch_beta(shape1 = 1, shape2 = 1)
}

