% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address-provider.R, R/available_locales.R,
%   R/color-provider.R, R/company-provider.R, R/file-provider.R,
%   R/internet-provider.R, R/jobs-provider.R, R/lorem-provider.R,
%   R/person-provider.R, R/phonenumbers-provider.R, R/useragent-provider.R
\docType{data}
\name{address_provider_locales}
\alias{address_provider_locales}
\alias{per_provider_locales}
\alias{color_provider_locales}
\alias{company_provider_locales}
\alias{file_provider_locales}
\alias{internet_provider_locales}
\alias{job_provider_locales}
\alias{lorem_provider_locales}
\alias{person_provider_locales}
\alias{phone_number_provider_locales}
\alias{useragent_provider_locales}
\title{per provider locale vectors}
\format{An object of class \code{character} of length 4.}
\usage{
address_provider_locales

color_provider_locales

company_provider_locales

file_provider_locales

internet_provider_locales

job_provider_locales

lorem_provider_locales

person_provider_locales

phone_number_provider_locales

useragent_provider_locales
}
\description{
vectors of allowed locales used for each provider of the
same name
}
\keyword{datasets}
