% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person-provider-lv_LV.R
\name{PersonProvider_lv_LV}
\alias{PersonProvider_lv_LV}
\title{Person Provider for Latvian (Latvia)}
\description{
person names methods
}
\details{
Note for female and male components that we fall back on generic
versions if the locale
doesn't provide a male/female version.
e.g., if no female first name we use first
name
}
\examples{
x <- PersonProvider_lv_LV$new()
x$locale
x$render()
x$first_name()
x$first_name_female()
x$first_name_male()
x$last_name()
x$last_name_female()
x$last_name_male()
}
\seealso{
Other lv: 
\code{\link{PhoneNumberProvider_lv_LV}},
\code{\link{latvian-language}}

Other LV: 
\code{\link{PhoneNumberProvider_lv_LV}}
}
\concept{LV}
\concept{lv}
\section{Super classes}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{\link[charlatan:PersonProvider]{charlatan::PersonProvider}} -> \code{PersonProvider_lv_LV}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PersonProvider_lv_LV-clone}{\code{PersonProvider_lv_LV$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="allowed_locales"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-allowed_locales'><code>charlatan::BaseProvider$allowed_locales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="print"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-print'><code>charlatan::BaseProvider$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="change_messy"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-change_messy'><code>charlatan::PersonProvider$change_messy()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="first_name"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-first_name'><code>charlatan::PersonProvider$first_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="first_name_female"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-first_name_female'><code>charlatan::PersonProvider$first_name_female()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="first_name_male"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-first_name_male'><code>charlatan::PersonProvider$first_name_male()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="initialize"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-initialize'><code>charlatan::PersonProvider$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="last_name"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-last_name'><code>charlatan::PersonProvider$last_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="last_name_female"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-last_name_female'><code>charlatan::PersonProvider$last_name_female()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="last_name_male"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-last_name_male'><code>charlatan::PersonProvider$last_name_male()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="messy_is_possible"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-messy_is_possible'><code>charlatan::PersonProvider$messy_is_possible()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="messy_switch"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-messy_switch'><code>charlatan::PersonProvider$messy_switch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="prefix"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-prefix'><code>charlatan::PersonProvider$prefix()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="prefix_female"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-prefix_female'><code>charlatan::PersonProvider$prefix_female()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="prefix_male"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-prefix_male'><code>charlatan::PersonProvider$prefix_male()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="render"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-render'><code>charlatan::PersonProvider$render()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="suffix"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-suffix'><code>charlatan::PersonProvider$suffix()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="suffix_female"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-suffix_female'><code>charlatan::PersonProvider$suffix_female()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="PersonProvider" data-id="suffix_male"><a href='../../charlatan/html/PersonProvider.html#method-PersonProvider-suffix_male'><code>charlatan::PersonProvider$suffix_male()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider_lv_LV-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider_lv_LV-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider_lv_LV$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
