% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonenumbers-provider.R
\docType{data}
\name{PhoneNumberProvider}
\alias{PhoneNumberProvider}
\alias{phone_number_provider_locales}
\title{PhoneNumberProvider}
\format{An object of class \code{character} of length 37.}
\usage{
phone_number_provider_locales
}
\arguments{
\item{locale}{(character) the locale to use. Run
\code{phone_number_provider_locales()} for locales supported (default: en_US)}
}
\description{
PhoneNumberProvider
}
\details{
\strong{Methods}
\itemize{
\item \code{render()} - Make a phone number
}
}
\examples{
z <- PhoneNumberProvider$new()
z$render()

PhoneNumberProvider$new(locale = "fr_FR")$render()
PhoneNumberProvider$new(locale = "sk_SK")$render()
}
\keyword{datasets}
\keyword{internal}
