% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence-provider.R
\docType{data}
\name{SequenceProvider}
\alias{SequenceProvider}
\title{SequenceProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SequenceProvider
}
\arguments{
\item{length}{(integer) length of sequence to create}
}
\description{
SequenceProvider
}
\examples{
z <- SequenceProvider$new()
z$render()
z$render(10)
z$render(100)
z$render(500)
}
\keyword{internal}

