% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\name{spanish-language}
\alias{spanish-language}
\title{Spanish Language}
\description{
Providers with the Spanish locale (es).
}
\seealso{
Other es: 
\code{\link{CompanyProvider_es_MX}},
\code{\link{PersonProvider_es_ES}},
\code{\link{PersonProvider_es_MX}},
\code{\link{PhoneNumberProvider_es_ES}},
\code{\link{PhoneNumberProvider_es_PE}}

Other languages: 
\code{\link{arabic-language}},
\code{\link{bosnian-language}},
\code{\link{bulgarian-language}},
\code{\link{chinese-language}},
\code{\link{croatian-language}},
\code{\link{czech-language}},
\code{\link{danish-language}},
\code{\link{dutch-language}},
\code{\link{english-language}},
\code{\link{farsi-language}},
\code{\link{finnish-language}},
\code{\link{french-language}},
\code{\link{german-language}},
\code{\link{greek-language}},
\code{\link{hebrew-language}},
\code{\link{hindi-language}},
\code{\link{hungarian-language}},
\code{\link{indonesian-language}},
\code{\link{italian-language}},
\code{\link{japanese-language}},
\code{\link{korean-language}},
\code{\link{latin-language}},
\code{\link{latvian-language}},
\code{\link{lithuanian-language}},
\code{\link{norwegian-language}},
\code{\link{polish-language}},
\code{\link{portuguese-language}},
\code{\link{russian-language}},
\code{\link{swedish-language}},
\code{\link{thai-language}},
\code{\link{turkish-language}},
\code{\link{ukrainian-language}}
}
\concept{es}
\concept{languages}
