% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssn-provider-all.R
\name{SSNProvider_nl_NL}
\alias{SSNProvider_nl_NL}
\title{SSNProvider the Netherlands}
\description{
methods for generating social security numbers
}
\examples{
z <- SSNProvider_nl_NL$new()
z$render()
}
\seealso{
Other nl: 
\code{\link{AddressProvider_nl_NL}},
\code{\link{ElementProvider_nl_NL}},
\code{\link{JobProvider_nl_NL}},
\code{\link{PersonProvider_nl_NL}},
\code{\link{PhoneNumberProvider_nl_BE}},
\code{\link{PhoneNumberProvider_nl_NL}},
\code{\link{dutch-language}}

Other NL: 
\code{\link{AddressProvider_nl_NL}},
\code{\link{ElementProvider_nl_NL}},
\code{\link{JobProvider_nl_NL}},
\code{\link{PersonProvider_nl_NL}},
\code{\link{PhoneNumberProvider_nl_NL}}
}
\concept{NL}
\concept{nl}
\section{Super classes}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{\link[charlatan:SSNProvider]{charlatan::SSNProvider}} -> \code{SSNProvider_nl_NL}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{locale}}{(character) the locale}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SSNProvider_nl_NL-render}{\code{SSNProvider_nl_NL$render()}}
\item \href{#method-SSNProvider_nl_NL-clone}{\code{SSNProvider_nl_NL$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="SSNProvider" data-id="allowed_locales"><a href='../../charlatan/html/SSNProvider.html#method-SSNProvider-allowed_locales'><code>charlatan::SSNProvider$allowed_locales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="SSNProvider" data-id="initialize"><a href='../../charlatan/html/SSNProvider.html#method-SSNProvider-initialize'><code>charlatan::SSNProvider$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSNProvider_nl_NL-render"></a>}}
\if{latex}{\out{\hypertarget{method-SSNProvider_nl_NL-render}{}}}
\subsection{Method \code{render()}}{
Make a SSN
Dutch SSN (BSN) is 9 digits that follow a certain proof (elfproef).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSNProvider_nl_NL$render()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SSNProvider_nl_NL-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SSNProvider_nl_NL-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSNProvider_nl_NL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
