% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{subclass}
\alias{subclass}
\title{Create Localized Provider}
\usage{
subclass(provider, locale = NULL)
}
\arguments{
\item{provider}{The name of the provider you want to create}

\item{locale}{Locale to use}
}
\value{
Localized provider
}
\description{
Create Localized Provider
}
\examples{
x <- subclass("AddressProvider")
}
