% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy}
\alias{taxonomy}
\alias{ch_taxonomic_genus}
\alias{ch_taxonomic_epithet}
\alias{ch_taxonomic_species}
\title{Create fake taxonomic names}
\usage{
ch_taxonomic_genus(n = 1)

ch_taxonomic_epithet(n = 1)

ch_taxonomic_species(n = 1)
}
\arguments{
\item{n}{(integer) number of things to get, any non-negative integer}
}
\description{
Create fake taxonomic names
}
\section{Names}{

Names were taken from Theplantlist. 500 genera names and 500
epithets were chosen at random from the set of 10,000 names in the
dataset in the \code{taxize} package. Theplantlist is, as it says on the
tin, composed of plant names - so these fake names are derived from
plant names if that matters to you. These may generate names that match
those of real taxa, but may not as well.
}

\section{Taxonomic authority}{

Randomly, the taxonomic authority is in parentheses - which represents
that the given authority was not the original authority.
}

\examples{
ch_taxonomic_genus()
ch_taxonomic_genus(10)
ch_taxonomic_genus(500)

ch_taxonomic_epithet()
ch_taxonomic_epithet(10)
ch_taxonomic_epithet(500)

ch_taxonomic_species()
ch_taxonomic_species(10)
ch_taxonomic_species(500)
}
