% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useragent-provider-all.R
\name{UserAgentProvider_en_US}
\alias{UserAgentProvider_en_US}
\title{UserAgentProvider for United States of America}
\description{
user agent methods
}
\examples{
(x <- UserAgentProvider_en_US$new())
x$locale
x$mac_processor()
x$linux_processor()
x$user_agent()
x$chrome()
x$firefox()
x$internet_explorer()
x$opera()
x$safari()
}
\seealso{
Other en: 
\code{\link{AddressProvider_en_GB}},
\code{\link{AddressProvider_en_NZ}},
\code{\link{ColorProvider_en_US}},
\code{\link{CompanyProvider_en_US}},
\code{\link{ElementProvider_en_US}},
\code{\link{InternetProvider_en_AU}},
\code{\link{InternetProvider_en_US}},
\code{\link{JobProvider_en_US}},
\code{\link{LoremProvider_en_US}},
\code{\link{PersonProvider_en_GB}},
\code{\link{PersonProvider_en_NZ}},
\code{\link{PersonProvider_en_US}},
\code{\link{PhoneNumberProvider_en_AU}},
\code{\link{PhoneNumberProvider_en_CA}},
\code{\link{PhoneNumberProvider_en_GB}},
\code{\link{PhoneNumberProvider_en_NZ}},
\code{\link{PhoneNumberProvider_en_US}},
\code{\link{PhoneNumberProvider_es_MX}},
\code{\link{PhoneNumberProvider_tr_TR}},
\code{\link{PhoneNumberProvider}},
\code{\link{SSNProvider_en_US}},
\code{\link{TaxonomyProvider_en_US}},
\code{\link{UserAgentProvider}},
\code{\link{english-language}}

Other US: 
\code{\link{ColorProvider_en_US}},
\code{\link{CompanyProvider_en_US}},
\code{\link{ElementProvider_en_US}},
\code{\link{InternetProvider_en_US}},
\code{\link{JobProvider_en_US}},
\code{\link{LoremProvider_en_US}},
\code{\link{PersonProvider_en_US}},
\code{\link{PhoneNumberProvider_en_US}},
\code{\link{SSNProvider_en_US}},
\code{\link{TaxonomyProvider_en_US}},
\code{\link{UserAgentProvider}}
}
\concept{US}
\concept{en}
\section{Super classes}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{\link[charlatan:UserAgentProvider]{charlatan::UserAgentProvider}} -> \code{UserAgentProvider_en_US}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{locale}}{(character) the locale}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-UserAgentProvider_en_US-clone}{\code{UserAgentProvider_en_US$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="UserAgentProvider" data-id="allowed_locales"><a href='../../charlatan/html/UserAgentProvider.html#method-UserAgentProvider-allowed_locales'><code>charlatan::UserAgentProvider$allowed_locales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="UserAgentProvider" data-id="chrome"><a href='../../charlatan/html/UserAgentProvider.html#method-UserAgentProvider-chrome'><code>charlatan::UserAgentProvider$chrome()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="UserAgentProvider" data-id="firefox"><a href='../../charlatan/html/UserAgentProvider.html#method-UserAgentProvider-firefox'><code>charlatan::UserAgentProvider$firefox()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="UserAgentProvider" data-id="initialize"><a href='../../charlatan/html/UserAgentProvider.html#method-UserAgentProvider-initialize'><code>charlatan::UserAgentProvider$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="UserAgentProvider" data-id="internet_explorer"><a href='../../charlatan/html/UserAgentProvider.html#method-UserAgentProvider-internet_explorer'><code>charlatan::UserAgentProvider$internet_explorer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="UserAgentProvider" data-id="linux_platform_token"><a href='../../charlatan/html/UserAgentProvider.html#method-UserAgentProvider-linux_platform_token'><code>charlatan::UserAgentProvider$linux_platform_token()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="UserAgentProvider" data-id="linux_processor"><a href='../../charlatan/html/UserAgentProvider.html#method-UserAgentProvider-linux_processor'><code>charlatan::UserAgentProvider$linux_processor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="UserAgentProvider" data-id="mac_platform_token"><a href='../../charlatan/html/UserAgentProvider.html#method-UserAgentProvider-mac_platform_token'><code>charlatan::UserAgentProvider$mac_platform_token()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="UserAgentProvider" data-id="mac_processor"><a href='../../charlatan/html/UserAgentProvider.html#method-UserAgentProvider-mac_processor'><code>charlatan::UserAgentProvider$mac_processor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="UserAgentProvider" data-id="opera"><a href='../../charlatan/html/UserAgentProvider.html#method-UserAgentProvider-opera'><code>charlatan::UserAgentProvider$opera()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="UserAgentProvider" data-id="safari"><a href='../../charlatan/html/UserAgentProvider.html#method-UserAgentProvider-safari'><code>charlatan::UserAgentProvider$safari()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="UserAgentProvider" data-id="user_agent"><a href='../../charlatan/html/UserAgentProvider.html#method-UserAgentProvider-user_agent'><code>charlatan::UserAgentProvider$user_agent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="UserAgentProvider" data-id="windows_platform_token"><a href='../../charlatan/html/UserAgentProvider.html#method-UserAgentProvider-windows_platform_token'><code>charlatan::UserAgentProvider$windows_platform_token()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-UserAgentProvider_en_US-clone"></a>}}
\if{latex}{\out{\hypertarget{method-UserAgentProvider_en_US-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{UserAgentProvider_en_US$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
