% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address-provider.R
\docType{data}
\name{AddressProvider}
\alias{AddressProvider}
\title{AddressProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AddressProvider
}
\description{
AddressProvider
}
\examples{
(z <- AddressProvider$new())
z$city_suffix()
z$street_suffix()
z$building_number()
z$city()
z$country()
z$country_code()
z$postcode()
}
\keyword{internal}

