% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address-provider.R
\docType{data}
\name{AddressProvider}
\alias{AddressProvider}
\alias{address_provider_locales}
\title{AddressProvider}
\format{An object of class \code{character} of length 2.}
\usage{
address_provider_locales
}
\arguments{
\item{locale}{(character) the locale to use. See
\code{address_provider_locales} for locales supported (default: en_US)}
}
\description{
AddressProvider
}
\details{
\strong{Methods}
\itemize{
\item \code{city_suffix()} - city suffix
\item \code{street_suffix()} - street suffix
\item \code{building_number()} - building number
\item \code{city()} - city
\item \code{street_name()} - street name
\item \code{street_address()} - street address
\item \code{address()} - address
\item \code{country()} - country
\item \code{country_code()} - country code
\item \code{postcode()} - postal code
}
}
\examples{
(z <- AddressProvider$new())
z$locale
z$city_suffix()
z$street_suffix()
z$building_number()
z$city()
z$country()
z$street_name()
z$street_address()
z$address()
z$country()
z$country_code()
z$postcode()

# en_GB
(z <- AddressProvider$new('en_GB'))
z$locale
z$locale_data
z$locale_data$postcode_sets
z$postcode
z$postcode()
z$street_name()
}
\keyword{datasets}
\keyword{internal}
