% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/address-provider.R
\docType{data}
\name{AddressProvider}
\alias{AddressProvider}
\title{AddressProvider}
\description{
AddressProvider
}
\details{
\strong{Methods}
  \describe{
   \item{\code{city_suffix}}{
     city suffix
   }
   \item{\code{street_suffix}}{
     street suffix
   }
   \item{\code{building_number}}{
     building number
   }
   \item{\code{city}}{
     city
   }
   \item{\code{country}}{
     country
   }
   \item{\code{country_code}}{
     country code
   }
   \item{\code{postcode}}{
     postal code
   }
 }
}
\examples{
(z <- AddressProvider$new())
z$city_suffix()
z$street_suffix()
z$building_number()
z$city()
z$country()
z$country_code()
z$postcode()
}
\keyword{internal}

