% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_locales.R
\docType{data}
\name{available_locales_df}
\alias{available_locales_df}
\title{Available locales}
\format{A data frame with 44 rows and 4 variables:
\describe{
\item{language}{language two letter code}
\item{country}{country two letter code}
\item{variant}{a variant code, if applicable}
\item{name}{official locale two letter code}
}}
\description{
A data.frame of locales available in \pkg{charlatan}
}
\seealso{
data.frame used in \code{\link[=charlatan_locales]{charlatan_locales()}}
}
\keyword{data}
