% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-provider.R
\name{BaseProvider}
\alias{BaseProvider}
\title{BaseProvider}
\description{
BaseProvider

BaseProvider
}
\details{
\strong{Methods}
\itemize{
\item \code{random_element(x)} - pick a random element from any input vector or list
\item \code{random_int(min = 0, max = 9999, size = 1)} - random integer
\item \code{random_digit()} - random integer between 0 and 9
\item \code{random_digit_not_zero()} - random integer between 1 and 9
\item \code{random_digit_or_empty()} - random integer between 0 and 9 or empty
character string
\item \code{random_digit_not_zero_or_empty()} - random integer between 1 and 9 or empty
character string
\item \code{random_letter()} - random letter
\item \code{numerify(text = '###')} - replace a template with numbers
\item \code{lexify(text = '????')} - replace a template with letters
\item \code{bothify(text = '## ??')} - both numerify and lexify together
\item \code{check_locale(x)} - check a locale to see if it exists, if not, stop with
error message
\item \code{randomize_nb_elements(x)} - Returns a random value near number.
\itemize{
\item param number: value to which the result must be near
\item param le: result must be lower or equal to number
\item param ge: result must be greater or equal to number
\item returns: a random int near number
}
}
}
\examples{
(x <- BaseProvider$new())

x$numerify("#\%\%asdf221?")
x$lexify("#\%\%asdf221?")
x$bothify("#\%\%asdf221?")

z <- PhoneNumberProvider$new()
x$numerify(z$render())

x$random_element(letters)
x$random_int()
x$random_digit()
x$random_digit_not_zero()
x$random_digit_or_empty()
x$random_digit_not_zero_or_empty()
x$random_letter()
x$check_locale("es_ES")
## fails
# x$check_locale("es_EQ")

x$randomize_nb_elements()
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-random_element}{\code{BaseProvider$random_element()}}
\item \href{#method-random_element_prob}{\code{BaseProvider$random_element_prob()}}
\item \href{#method-random_int}{\code{BaseProvider$random_int()}}
\item \href{#method-random_digit}{\code{BaseProvider$random_digit()}}
\item \href{#method-random_digit_not_zero}{\code{BaseProvider$random_digit_not_zero()}}
\item \href{#method-random_digit_or_empty}{\code{BaseProvider$random_digit_or_empty()}}
\item \href{#method-random_digit_not_zero_or_empty}{\code{BaseProvider$random_digit_not_zero_or_empty()}}
\item \href{#method-random_letter}{\code{BaseProvider$random_letter()}}
\item \href{#method-numerify}{\code{BaseProvider$numerify()}}
\item \href{#method-lexify}{\code{BaseProvider$lexify()}}
\item \href{#method-bothify}{\code{BaseProvider$bothify()}}
\item \href{#method-check_locale}{\code{BaseProvider$check_locale()}}
\item \href{#method-prov_avail_locales}{\code{BaseProvider$prov_avail_locales()}}
\item \href{#method-randomize_nb_elements}{\code{BaseProvider$randomize_nb_elements()}}
\item \href{#method-clone}{\code{BaseProvider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_element"></a>}}
\subsection{Method \code{random_element()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$random_element(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_element_prob"></a>}}
\subsection{Method \code{random_element_prob()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$random_element_prob(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_int"></a>}}
\subsection{Method \code{random_int()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$random_int(min = 0, max = 9999, size = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_digit"></a>}}
\subsection{Method \code{random_digit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$random_digit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_digit_not_zero"></a>}}
\subsection{Method \code{random_digit_not_zero()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$random_digit_not_zero()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_digit_or_empty"></a>}}
\subsection{Method \code{random_digit_or_empty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$random_digit_or_empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_digit_not_zero_or_empty"></a>}}
\subsection{Method \code{random_digit_not_zero_or_empty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$random_digit_not_zero_or_empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-random_letter"></a>}}
\subsection{Method \code{random_letter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$random_letter()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-numerify"></a>}}
\subsection{Method \code{numerify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$numerify(text = "###")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lexify"></a>}}
\subsection{Method \code{lexify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$lexify(text = "????")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bothify"></a>}}
\subsection{Method \code{bothify()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$bothify(text = "## ??")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_locale"></a>}}
\subsection{Method \code{check_locale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$check_locale(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prov_avail_locales"></a>}}
\subsection{Method \code{prov_avail_locales()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$prov_avail_locales(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-randomize_nb_elements"></a>}}
\subsection{Method \code{randomize_nb_elements()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$randomize_nb_elements(
  number = 10,
  le = FALSE,
  ge = FALSE,
  min = NULL,
  max = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
