% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-provider.R
\name{BaseProvider}
\alias{BaseProvider}
\title{BaseProvider}
\description{
BaseProvider

BaseProvider
}
\note{
You cannot instantiate the Parent providers. You must use one of
the localized one.
}
\examples{
(x <- BaseProvider$new())

x$numerify("#\%\%asdf221?")
x$lexify("#\%\%asdf221?")
x$bothify("#\%\%asdf221?")
}
\section{Super class}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{BaseProvider}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{locale}}{(character) locale of this Provider.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BaseProvider-check_locale}{\code{BaseProvider$check_locale()}}
\item \href{#method-BaseProvider-allowed_locales}{\code{BaseProvider$allowed_locales()}}
\item \href{#method-BaseProvider-new}{\code{BaseProvider$new()}}
\item \href{#method-BaseProvider-print}{\code{BaseProvider$print()}}
\item \href{#method-BaseProvider-clone}{\code{BaseProvider$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseProvider-check_locale"></a>}}
\if{latex}{\out{\hypertarget{method-BaseProvider-check_locale}{}}}
\subsection{Method \code{check_locale()}}{
check a locale to see if it exists, if not, stop with
error message
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$check_locale(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{a locale name, e.g, 'bg_BG'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
returns nothing if locale is supported; stops w/ message if not
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseProvider-allowed_locales"></a>}}
\if{latex}{\out{\hypertarget{method-BaseProvider-allowed_locales}{}}}
\subsection{Method \code{allowed_locales()}}{
fetch the allowed locales for this provider
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$allowed_locales()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseProvider-new"></a>}}
\if{latex}{\out{\hypertarget{method-BaseProvider-new}{}}}
\subsection{Method \code{new()}}{
Create a new Provider object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseProvider-print"></a>}}
\if{latex}{\out{\hypertarget{method-BaseProvider-print}{}}}
\subsection{Method \code{print()}}{
Print method for provider
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored by this method}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BaseProvider-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BaseProvider-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BaseProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
