% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-provider.R
\docType{data}
\name{BaseProvider}
\alias{BaseProvider}
\title{BaseProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BaseProvider
}
\description{
BaseProvider
}
\examples{
\donttest{
(x <- BaseProvider$new())

x$numerify("#\%\%asdf221?")
x$lexify("#\%\%asdf221?")
x$bothify("#\%\%asdf221?")

z <- PhoneNumberProvider$new()
x$numerify(z$random_format())

x$random_letter()
}
}
\keyword{internal}

