% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-provider.R
\docType{data}
\name{BaseProvider}
\alias{BaseProvider}
\title{BaseProvider}
\description{
BaseProvider
}
\examples{
(x <- BaseProvider$new())

x$numerify("#\%\%asdf221?")
x$lexify("#\%\%asdf221?")
x$bothify("#\%\%asdf221?")

z <- PhoneNumberProvider$new()
x$numerify(z$render())

x$random_element(letters)
x$random_int()
x$random_digit()
x$random_digit_not_zero()
x$random_digit_or_empty()
x$random_digit_not_zero_or_empty()
x$random_letter()
x$check_locale("es_ES")
## fails
# x$check_locale("es_EQ")

x$randomize_nb_elements()
}
\keyword{internal}
