% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-provider.R
\docType{data}
\name{BaseProvider}
\alias{BaseProvider}
\title{BaseProvider}
\description{
BaseProvider
}
\details{
\strong{Methods}
  \describe{
   \item{\code{random_element}}{
     wrapper around \code{\link{sample}}
   }
   \item{\code{random_int}}{
     random integer
   }
   \item{\code{random_digit}}{
     random integer between 0 and 9
   }
   \item{\code{random_digit_not_null}}{
     random integer between 1 and 9
   }
   \item{\code{random_digit_or_empty}}{
     random integer between 0 and 9 or empty character string
   }
   \item{\code{random_digit_not_null_or_empty}}{
     random integer between 1 and 9 or empty character string
   }
   \item{\code{random_letter}}{
     random letter
   }
   \item{\code{numerify}}{
     replace a template with numbers
   }
   \item{\code{lexify}}{
     replace a template with letters
   }
   \item{\code{bothify}}{
     both numerify and lexify together
   }
   \item{\code{check_locale}}{
     check a locale to see if it exists, if not, stop with error message
   }
 }
}
\examples{
(x <- BaseProvider$new())

x$numerify("#\%\%asdf221?")
x$lexify("#\%\%asdf221?")
x$bothify("#\%\%asdf221?")

z <- PhoneNumberProvider$new()
x$numerify(z$render())

x$random_element(letters)
x$random_int()
x$random_digit()
x$random_digit_not_null()
x$random_digit_or_empty()
x$random_digit_not_null_or_empty()
x$random_letter()
x$check_locale("es_ES")
## fails
# x$check_locale("es_EQ")
}
\keyword{internal}

