% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{ch_color}
\alias{ch_color}
\alias{ch_color_name}
\alias{ch_safe_color_name}
\alias{ch_hex_color}
\alias{ch_safe_hex_color}
\alias{ch_rgb_color}
\alias{ch_rgb_css_color}
\title{Create fake colors}
\usage{
ch_color_name(n = 1, locale = NULL)

ch_safe_color_name(n = 1)

ch_hex_color(n = 1)

ch_safe_hex_color(n = 1)

ch_rgb_color(n = 1)

ch_rgb_css_color(n = 1)
}
\arguments{
\item{n}{(integer) number of things to get, any positive integer > 1}

\item{locale}{(character) the locale to use. options: en_US (default),
or uk_UA. Only affects \code{ch_color_name} function}
}
\description{
Create fake colors
}
\examples{
ch_color_name()
ch_color_name(10)
ch_color_name(500)

ch_safe_color_name()
ch_safe_color_name(10)

ch_hex_color()
ch_hex_color(10)

ch_safe_hex_color()
ch_safe_hex_color(10)

ch_rgb_color()
ch_rgb_color(10)

ch_rgb_css_color()
ch_rgb_css_color(10)

ch_color_name(locale = "uk_UA")
ch_color_name(n = 10, locale = "uk_UA")
}
