% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/company.R
\name{ch_company}
\alias{ch_company}
\title{Create fake company names and other company bits}
\usage{
ch_company(n = 1, locale = NULL)
}
\arguments{
\item{n}{(integer) number of things to get, any non-negative integer}

\item{locale}{(character) the locale to use. See
\code{company()$allowed_locales()} for locales supported.}
}
\description{
Create fake company names and other company bits
}
\examples{
ch_company()
ch_company(10)
ch_company(500)

ch_company(locale = "fr_FR", n = 10)
ch_company(locale = "cs_CZ", n = 10)
ch_company(locale = "es_MX", n = 10)
ch_company(locale = "hr_HR", n = 10)
}
\seealso{
\link{CompanyProvider}
}
