% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{ch_currency}
\alias{ch_currency}
\title{Create fake currencies}
\usage{
ch_currency(n = 1)
}
\arguments{
\item{n}{(integer) number of things to get, any non-negative integer}
}
\description{
Create fake currencies
}
\examples{
ch_currency()
ch_currency(10)
ch_currency(500)
}
