% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{ch_job}
\alias{ch_job}
\title{Create fake jobs}
\usage{
ch_job(n = 1)
}
\arguments{
\item{n}{(integer) number of things to get, 1 to inifinity}
}
\description{
Create fake jobs
}
\examples{
ch_job()
ch_job(10)
ch_job(500)
}

