% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{ch_job}
\alias{ch_job}
\title{Create fake jobs}
\usage{
ch_job(n = 1, locale = NULL)
}
\arguments{
\item{n}{(integer) number of things to get, 1 to inifinity}

\item{locale}{(character) the locale to use. only options right now are
en_US, fr_FR, hr_FR, more coming. default: en_US}
}
\description{
Create fake jobs
}
\examples{
ch_job()
ch_job(10)
ch_job(500)

ch_job(locale = "fr_FR", n = 10)
ch_job(locale = "hr_HR", n = 10)
}

