% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{ch_job}
\alias{ch_job}
\title{Create fake jobs}
\usage{
ch_job(n = 1, locale = NULL)
}
\arguments{
\item{n}{(integer) number of things to get, any non-negative integer}

\item{locale}{(character) the locale to use. options: en_US (default),
fr_FR, fr_CH, hr_HR, fa_IR, pl_PL, ru_RU, uk_UA, zh_TW.}
}
\description{
Create fake jobs
}
\examples{
ch_job()
ch_job(10)
ch_job(500)

ch_job(locale = "fr_FR", n = 10)
ch_job(locale = "fr_CH", n = 10)
ch_job(locale = "hr_HR", n = 10)
ch_job(locale = "fa_IR", n = 10)
ch_job(locale = "pl_PL", n = 10)
ch_job(locale = "ru_RU", n = 10)
ch_job(locale = "uk_UA", n = 10)
ch_job(locale = "zh_TW", n = 10)
}
