% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{ch_name}
\alias{ch_name}
\title{Create fake person names}
\usage{
ch_name(n = 1, locale = NULL)
}
\arguments{
\item{n}{(integer) number of things to get, any positive integer > 1}

\item{locale}{(character) the locale to use. options: en_US (default),
fr_FR, fr_CH, hr_HR, fa_IR, pl_PL, ru_RU, uk_UA, zh_TW.}
}
\description{
Create fake person names
}
\examples{
ch_name()
ch_name(10)
ch_name(500)

ch_name(locale = "fr_FR", n = 10)
ch_name(locale = "fr_CH", n = 10)
ch_name(locale = "fa_IR", n = 10)
ch_name(locale = "fi_FI", n = 10)
}
