% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{ch_name}
\alias{ch_name}
\title{Create fake person names}
\usage{
ch_name(n = 1)
}
\arguments{
\item{n}{(integer) number of things to get, 1 to inifinity}
}
\description{
Create fake person names
}
\examples{
ch_name()
ch_name(10)
ch_name(500)
}

