% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone_number.R
\name{ch_phone_number}
\alias{ch_phone_number}
\title{Create fake phone numbers}
\usage{
ch_phone_number(n = 1, locale = NULL)
}
\arguments{
\item{n}{(integer) number of things to get, any non-negative integer}

\item{locale}{(character) the locale to use. options: en_US (default),
en_US, es_ES, es_MX, en_GB, en_CA, el_GR, dk_DK, de_DE, cs_CZ, bs_BA,
bg_BG, fa_IR, fi_FI, fr_CH, fr_FR, hi_IN, hr_HR, hu_HU, it_IT, ja_JP, ko_KR,
it_LT, lv_LV, ne_NP, nl_BE, nl_NL, no_NO, pl_PL, pt_BR, pt_PT, ru_RU,
sk_SK, sl_SL, sv_SE, tr_TR, uk_UA, zh_TW}
}
\description{
Create fake phone numbers
}
\examples{
ch_phone_number()
ch_phone_number(10)
ch_phone_number(500)

# locales
ch_phone_number(locale = "fr_FR")
ch_phone_number(locale = "uk_UA")
ch_phone_number(locale = "en_CA")
ch_phone_number(locale = "lv_LV")
}
