% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone_number.R
\name{ch_phone_number}
\alias{ch_phone_number}
\title{Create fake phone numbers}
\usage{
ch_phone_number(n = 1, locale = NULL)
}
\arguments{
\item{n}{(integer) number of things to get, any non-negative integer}

\item{locale}{(character) the locale to use. See
\code{phone_number_provider_locales} for locales supported (default: en_US)}
}
\description{
Create fake phone numbers
}
\examples{
ch_phone_number()
ch_phone_number(10)
ch_phone_number(500)

# locales
ch_phone_number(locale = "fr_FR")
ch_phone_number(locale = "uk_UA")
ch_phone_number(locale = "en_CA")
ch_phone_number(locale = "lv_LV")
}
