% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phone_number.R
\name{ch_phone_number}
\alias{ch_phone_number}
\title{Create fake phone numbers}
\usage{
ch_phone_number()
}
\description{
Create fake phone numbers
}
\examples{
ch_phone_number()
replicate(10, ch_phone_number())
replicate(500, ch_phone_number())
}

