% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlatan-package.R
\docType{package}
\name{charlatan-package}
\alias{charlatan}
\alias{charlatan-package}
\title{charlatan}
\description{
charlatan
}
\section{Package API}{

\itemize{
 \item
}
}
\examples{
# generate individual types of data
ch_name()
ch_phone_number()
ch_job()

# generate a data.frame
ch_generate()

# one interface to all data types - generate the class first
#  reports the locale to be used, can change optionally
(x <- fraudster())
x$job()
x$name()
x$color_name()
x$hex_color()

# low level interfaces to "data providers"
# these are exported by hidden from package man page
# as most users will likely not interact with these
x <- ColorProvider$new()
x$color_name()
x$hex_color()
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

