% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_locales.R
\name{charlatan_locales}
\alias{charlatan_locales}
\title{Available locales}
\usage{
charlatan_locales()
}
\value{
a data.frame of the available locales in this package.
Not all functions support all locales. Check the docs for each one
to see what locales they support.

You can find out more about each locale by running your locale
though \code{stringi::stri_locale_info()}
}
\description{
Available locales
}
\examples{
charlatan_locales()
}
