% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlatan_settings.R
\name{charlatan_settings}
\alias{charlatan_settings}
\title{charlatan settings}
\usage{
charlatan_settings(messy = NULL)
}
\arguments{
\item{messy}{(logical) make some messy data. Default: \code{NULL}}
}
\description{
charlatan settings
}
\section{More deets}{

\itemize{
 \item \code{messy} - When \code{FALSE}, nothing is different from normal.
 When \code{TRUE}, we select incorrect/wrong values with probability X.
 Messy mode is only available for \strong{en-US} for now, and only for
 some data types. The default setting is \code{NULL}, meaning it is
 ignored.
}
}

\examples{
charlatan_settings()
charlatan_settings(messy = TRUE)
charlatan_settings(messy = FALSE)

# with PersonProvider - overrides local messy param in all cases
x <- PersonProvider$new()
x$messy
charlatan_settings(messy = TRUE)
x <- PersonProvider$new()
x$messy
}
