% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checksum_util}
\alias{checksum_util}
\title{generalized util function for sequence multiplication}
\usage{
checksum_util(vector, multiplicationvector)
}
\description{
Util function for ISBN, ean, SSN providers.
for cases such as \code{sum(x[4]*5 + x[3]*4 + x[2]*3 + x[1] *2)}
}
\keyword{internal}
