% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-provider.R
\docType{data}
\name{ColorProvider}
\alias{ColorProvider}
\title{ColorProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ColorProvider
}
\description{
ColorProvider
}
\examples{
\donttest{
x <- ColorProvider$new()
x$locale
x$color_name()
x$safe_color_name()
x$hex_color()
x$safe_hex_color()
x$rgb_color()
x$rgb_css_color()

x <- ColorProvider$new(locale = "uk_UA")
x$locale
x$color_name()
}
}
\keyword{internal}

