% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-provider.R
\name{ColorProvider}
\alias{ColorProvider}
\title{ColorProvider}
\description{
methods for colors
}
\examples{
x <- ColorProvider$new()
x$locale
x$color_name()
x$safe_color_name()
x$hex_color()
x$safe_hex_color()
x$rgb_color()
x$rgb_css_color()

x <- ColorProvider$new(locale = "uk_UA")
x$locale
x$color_name()
x$safe_color_name()
}
\keyword{internal}
\section{Super class}{
\code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{ColorProvider}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{locale}}{(character) xxx}

\item{\code{all_colors}}{(character) xxx}

\item{\code{safe_colors}}{(character) xxx}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ColorProvider-allowed_locales}{\code{ColorProvider$allowed_locales()}}
\item \href{#method-ColorProvider-new}{\code{ColorProvider$new()}}
\item \href{#method-ColorProvider-color_name}{\code{ColorProvider$color_name()}}
\item \href{#method-ColorProvider-safe_color_name}{\code{ColorProvider$safe_color_name()}}
\item \href{#method-ColorProvider-hex_color}{\code{ColorProvider$hex_color()}}
\item \href{#method-ColorProvider-safe_hex_color}{\code{ColorProvider$safe_hex_color()}}
\item \href{#method-ColorProvider-rgb_color}{\code{ColorProvider$rgb_color()}}
\item \href{#method-ColorProvider-rgb_css_color}{\code{ColorProvider$rgb_css_color()}}
\item \href{#method-ColorProvider-clone}{\code{ColorProvider$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="bothify"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-bothify'><code>charlatan::BaseProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="lexify"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-lexify'><code>charlatan::BaseProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="numerify"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-numerify'><code>charlatan::BaseProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_digit"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-random_digit'><code>charlatan::BaseProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-random_digit_not_zero'><code>charlatan::BaseProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-random_digit_not_zero_or_empty'><code>charlatan::BaseProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-random_digit_or_empty'><code>charlatan::BaseProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_element"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-random_element'><code>charlatan::BaseProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_element_prob"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-random_element_prob'><code>charlatan::BaseProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_int"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-random_int'><code>charlatan::BaseProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_letter"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-random_letter'><code>charlatan::BaseProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-randomize_nb_elements'><code>charlatan::BaseProvider$randomize_nb_elements()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-allowed_locales"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-allowed_locales}{}}}
\subsection{Method \code{allowed_locales()}}{
fetch the allowed locales for this provider
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$allowed_locales()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-new"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ColorProvider} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$new(locale = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{locale}}{(character) the locale to use. See
\verb{$allowed_locales()} for locales supported (default: en_US)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ColorProvider} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-color_name"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-color_name}{}}}
\subsection{Method \code{color_name()}}{
color name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$color_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-safe_color_name"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-safe_color_name}{}}}
\subsection{Method \code{safe_color_name()}}{
safe color name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$safe_color_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-hex_color"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-hex_color}{}}}
\subsection{Method \code{hex_color()}}{
hex color
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$hex_color()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-safe_hex_color"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-safe_hex_color}{}}}
\subsection{Method \code{safe_hex_color()}}{
safe hex color
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$safe_hex_color()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-rgb_color"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-rgb_color}{}}}
\subsection{Method \code{rgb_color()}}{
RGB color
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$rgb_color()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-rgb_css_color"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-rgb_css_color}{}}}
\subsection{Method \code{rgb_css_color()}}{
RGB CSS color
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$rgb_css_color()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ColorProvider-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ColorProvider-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ColorProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
