% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate-provider.R
\docType{data}
\name{CoordinateProvider}
\alias{CoordinateProvider}
\title{CoordinateProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CoordinateProvider
}
\description{
CoordinateProvider
}
\examples{
\donttest{
z <- CoordinateProvider$new()
z$lon()
z$lat()
z$position()
z$position(bbox = c(-120, 30, -110, 60))
}
}
\keyword{internal}

