% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate-provider.R
\docType{data}
\name{CoordinateProvider}
\alias{CoordinateProvider}
\title{CoordinateProvider}
\description{
CoordinateProvider
}
\details{
\strong{Methods}
\describe{
\item{\code{lat()}}{
a latitude value
}
\item{\code{lon()}}{
a longitude value
}
\item{\code{position()}}{
a position, of form \code{[longitude,latitude]}
\itemize{
\item bbox: optionally, specify a bounding box for the position
to be in, of the form \code{[west,south,east,north]} - checks that the
bbox has valid values for lat and long
}
}
}
}
\examples{
z <- CoordinateProvider$new()
z$lon()
z$lat()
z$position()
z$position(bbox = c(-120, 30, -110, 60))
}
\keyword{internal}
