% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency-provider.R
\docType{data}
\name{CurrencyProvider}
\alias{CurrencyProvider}
\title{CurrencyProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
CurrencyProvider
}
\description{
CurrencyProvider
}
\examples{
z <- CurrencyProvider$new()
z$render()
}
\keyword{internal}

