% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency-provider.R
\docType{data}
\name{CurrencyProvider}
\alias{CurrencyProvider}
\title{CurrencyProvider}
\description{
CurrencyProvider
}
\details{
\strong{Methods}
  \describe{
    \item{\code{render()}}{
      a currency, character string length 3
    }
  }
}
\examples{
z <- CurrencyProvider$new()
z$render()
}
\keyword{internal}

