% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime-provider.R
\docType{data}
\name{DateTimeProvider}
\alias{DateTimeProvider}
\title{DateTimeProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DateTimeProvider
}
\description{
DateTimeProvider
}
\examples{
\donttest{
z <- DateTimeProvider$new()
z$countries
z$centuries
z$timezone()
z$unix_time()
z$date_time()
z$year()
}
}
\keyword{internal}

