% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime-provider.R
\docType{data}
\name{DateTimeProvider}
\alias{DateTimeProvider}
\title{DateTimeProvider}
\description{
DateTimeProvider
}
\details{
\strong{Methods}
  \describe{
   \item{\code{unix_time}}{
     generate a unix time
   }
   \item{\code{date}}{
     generate a date
   }
   \item{\code{date_time}}{
     generate a date time
   }
   \item{\code{date_time_ad}}{
     generate a date time
   }
   \item{\code{iso8601}}{
     generate a iso8601 format date
   }
   \item{\code{year}}{
     generate a year
   }
   \item{\code{century}}{
     generate a century
   }
   \item{\code{timezone}}{
     generate a timezone
   }
 }
}
\examples{
z <- DateTimeProvider$new()
z$countries
z$centuries
z$century()
z$timezone()
z$unix_time()
z$date_time()
z$year()
}
\keyword{internal}
