% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi-provider.R
\docType{data}
\name{DOIProvider}
\alias{DOIProvider}
\title{DOIProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
DOIProvider
}
\description{
DOIProvider
}
\examples{
z <- DOIProvider$new()
z$render()
}
\keyword{internal}

