% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi-provider.R
\docType{data}
\name{DOIProvider}
\alias{DOIProvider}
\title{DOIProvider}
\description{
DOIProvider
}
\details{
\strong{Methods}
  \describe{
    \item{\code{render()}}{
      Make a DOI
    }
  }
}
\examples{
(z <- DOIProvider$new())
z$render()
}
\keyword{internal}
