% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-provider.R
\docType{data}
\name{ElementProvider}
\alias{ElementProvider}
\title{ElementProvider}
\description{
ElementProvider
}
\details{
\strong{Methods}
\describe{
\item{\code{symbol()}}{
Get a symbol
}
\item{\code{element()}}{
Get an element
}
}

Data from Wikipedia at
\url{https://en.wikipedia.org/wiki/Chemical_element}
}
\examples{
z <- ElementProvider$new()
z$symbol()
z$element()
}
\keyword{internal}
