% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/element-provider.R
\name{ElementProvider}
\alias{ElementProvider}
\title{ElementProvider}
\description{
chemical elements methods
}
\details{
Data from Wikipedia at
\url{https://en.wikipedia.org/wiki/Chemical_element}
}
\keyword{internal}
\section{Super classes}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{ElementProvider}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{locale}}{(character) xxx}

\item{\code{symbolvec}}{symbols for elements}

\item{\code{elementvec}}{vector of element names}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ElementProvider-elements}{\code{ElementProvider$elements()}}
\item \href{#method-ElementProvider-new}{\code{ElementProvider$new()}}
\item \href{#method-ElementProvider-symbol}{\code{ElementProvider$symbol()}}
\item \href{#method-ElementProvider-element}{\code{ElementProvider$element()}}
\item \href{#method-ElementProvider-symbol_by_number}{\code{ElementProvider$symbol_by_number()}}
\item \href{#method-ElementProvider-element_by_number}{\code{ElementProvider$element_by_number()}}
\item \href{#method-ElementProvider-clone}{\code{ElementProvider$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="allowed_locales"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-allowed_locales'><code>charlatan::BaseProvider$allowed_locales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementProvider-elements"></a>}}
\if{latex}{\out{\hypertarget{method-ElementProvider-elements}{}}}
\subsection{Method \code{elements()}}{
dataframe of symbols and elements
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementProvider$elements()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementProvider-new"></a>}}
\if{latex}{\out{\hypertarget{method-ElementProvider-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ElementProvider} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementProvider$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{ElementProvider} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementProvider-symbol"></a>}}
\if{latex}{\out{\hypertarget{method-ElementProvider-symbol}{}}}
\subsection{Method \code{symbol()}}{
Get a symbol
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementProvider$symbol()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementProvider-element"></a>}}
\if{latex}{\out{\hypertarget{method-ElementProvider-element}{}}}
\subsection{Method \code{element()}}{
Get an element
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementProvider$element()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementProvider-symbol_by_number"></a>}}
\if{latex}{\out{\hypertarget{method-ElementProvider-symbol_by_number}{}}}
\subsection{Method \code{symbol_by_number()}}{
symbol by number
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementProvider$symbol_by_number(number)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{number}}{retrieve symbol of element by atomic number.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementProvider-element_by_number"></a>}}
\if{latex}{\out{\hypertarget{method-ElementProvider-element_by_number}{}}}
\subsection{Method \code{element_by_number()}}{
element by number
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementProvider$element_by_number(number)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{number}}{retrieve element by atomic number.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ElementProvider-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ElementProvider-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ElementProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
