% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-provider.R
\docType{data}
\name{FileProvider}
\alias{FileProvider}
\alias{file_provider_locales}
\title{FileProvider}
\format{An object of class \code{character} of length 1.}
\usage{
file_provider_locales
}
\arguments{
\item{locale}{(character) the locale to use. Run
\code{file_provider_locales()} for locales supported (default: en_US)}
}
\description{
FileProvider
}
\details{
\strong{Methods}
\itemize{
\item \code{mime_type(category)} - mime type
\item \code{file_name(category(NULL, extension)} - file name
\item \code{file_extension(category)} - file extension
\item \code{file_path(depth(1, category(NULL, extension)} - file path
}
}
\examples{
(x <- FileProvider$new())
x$locale 
x$mime_type()
x$file_extension()
x$file_name()
x$file_path()
x$file_path(depth = 2)
x$file_path(depth = 3)
x$file_path(depth = 6)
}
\keyword{datasets}
\keyword{internal}
