% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internet-provider.R
\name{InternetProvider}
\alias{InternetProvider}
\title{InternetProvider}
\description{
internet methods, e.g., email addresses, domain names
}
\note{
Note that if a locale you set doesn't have a locale specific set
of data for \link{PersonProvider} or \link{CompanyProvider} we fall back to
\code{en_US}
}
\examples{
(x <- InternetProvider$new())
x$locale

# uri/url/tld/etc.
x$tld()
x$slug()
x$domain_word()
x$domain_name()
x$domain_name(levels = 2)
x$domain_name(levels = 3)
x$domain_name(levels = 10)
## url's
x$url()
x$url(schemes = c('hbbp', 'hggp'))
x$image_url()
## uri's
x$uri()
x$uri_page()
x$uri_extension()
x$uri_path()
x$uri_path(deep = 1)
x$uri_path(deep = 2)
x$uri_path(deep = 3)
x$uri_path(deep = 4)

# user name
x$user_name()

# emails
x$email()
x$safe_email()
x$free_email()
x$company_email()
x$free_email_domain()
x$ascii_email()
x$ascii_safe_email()
x$ascii_free_email()
x$ascii_company_email()

# addresses, mac, ipv4
x$mac_address()
if (requireNamespace("iptools", quietly=TRUE)) {
  x$ipv4()
}

# different locales
(x <- InternetProvider$new(locale = "en_AU"))
x$locale
x$tld()
x$email()
x$free_email_domain()

(x <- InternetProvider$new(locale = "de_DE"))
x$locale
x$tld()
x$uri()
x$email()
x$ascii_email()

(x <- InternetProvider$new(locale = "bg_BG"))
x$locale
x$tld()
x$uri()
x$url()
x$user_name()
x$email()
x$ascii_email()

(x <- InternetProvider$new(locale = "cs_CZ"))
x$url()
x$user_name()
x$email()

(x <- InternetProvider$new(locale = "en_NZ"))
x$free_email_domain()
x$tld()

(x <- InternetProvider$new(locale = "fa_IR"))
x$url()

(x <- InternetProvider$new(locale = "fr_FR"))
x$url()
x$user_name()
x$email()

(x <- InternetProvider$new(locale = "hr_HR"))
x$url()
x$user_name()
x$email()

# convert a string to ascii with stringi pkg
if (requireNamespace("stringi", quietly=TRUE)) {
  x$to_ascii("anï")
}
}
\keyword{internal}
\section{Super class}{
\code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{InternetProvider}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{locale}}{(character) the locale}

\item{\code{safe_email_tlds}}{(character) safe email tlds}

\item{\code{free_email_domains}}{(character) free email tlds}

\item{\code{tlds}}{(character) tlds}

\item{\code{uri_pages}}{(character) uri pages}

\item{\code{uri_paths}}{(character) uri paths}

\item{\code{uri_extensions}}{(character) uri extensions}

\item{\code{user_name_formats}}{(character) user name formats}

\item{\code{email_formats}}{(character) email formats}

\item{\code{url_formats}}{(character) url formats}

\item{\code{uri_formats}}{(character) uri formats}

\item{\code{image_placeholder_services}}{(character) image uri formats}

\item{\code{replacements}}{(list) a list}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-allowed_locales}{\code{InternetProvider$allowed_locales()}}
\item \href{#method-new}{\code{InternetProvider$new()}}
\item \href{#method-to_ascii}{\code{InternetProvider$to_ascii()}}
\item \href{#method-email}{\code{InternetProvider$email()}}
\item \href{#method-safe_email}{\code{InternetProvider$safe_email()}}
\item \href{#method-free_email}{\code{InternetProvider$free_email()}}
\item \href{#method-company_email}{\code{InternetProvider$company_email()}}
\item \href{#method-ascii_email}{\code{InternetProvider$ascii_email()}}
\item \href{#method-ascii_safe_email}{\code{InternetProvider$ascii_safe_email()}}
\item \href{#method-ascii_free_email}{\code{InternetProvider$ascii_free_email()}}
\item \href{#method-ascii_company_email}{\code{InternetProvider$ascii_company_email()}}
\item \href{#method-user_name}{\code{InternetProvider$user_name()}}
\item \href{#method-tld}{\code{InternetProvider$tld()}}
\item \href{#method-free_email_domain}{\code{InternetProvider$free_email_domain()}}
\item \href{#method-url}{\code{InternetProvider$url()}}
\item \href{#method-domain_name}{\code{InternetProvider$domain_name()}}
\item \href{#method-domain_word}{\code{InternetProvider$domain_word()}}
\item \href{#method-ipv4}{\code{InternetProvider$ipv4()}}
\item \href{#method-mac_address}{\code{InternetProvider$mac_address()}}
\item \href{#method-uri_page}{\code{InternetProvider$uri_page()}}
\item \href{#method-uri_path}{\code{InternetProvider$uri_path()}}
\item \href{#method-uri_extension}{\code{InternetProvider$uri_extension()}}
\item \href{#method-uri}{\code{InternetProvider$uri()}}
\item \href{#method-slug}{\code{InternetProvider$slug()}}
\item \href{#method-image_url}{\code{InternetProvider$image_url()}}
\item \href{#method-clone}{\code{InternetProvider$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="bothify">}\href{../../charlatan/html/BaseProvider.html#method-bothify}{\code{charlatan::BaseProvider$bothify()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale">}\href{../../charlatan/html/BaseProvider.html#method-check_locale}{\code{charlatan::BaseProvider$check_locale()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="lexify">}\href{../../charlatan/html/BaseProvider.html#method-lexify}{\code{charlatan::BaseProvider$lexify()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="numerify">}\href{../../charlatan/html/BaseProvider.html#method-numerify}{\code{charlatan::BaseProvider$numerify()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_digit">}\href{../../charlatan/html/BaseProvider.html#method-random_digit}{\code{charlatan::BaseProvider$random_digit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_digit_not_zero">}\href{../../charlatan/html/BaseProvider.html#method-random_digit_not_zero}{\code{charlatan::BaseProvider$random_digit_not_zero()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_digit_not_zero_or_empty">}\href{../../charlatan/html/BaseProvider.html#method-random_digit_not_zero_or_empty}{\code{charlatan::BaseProvider$random_digit_not_zero_or_empty()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_digit_or_empty">}\href{../../charlatan/html/BaseProvider.html#method-random_digit_or_empty}{\code{charlatan::BaseProvider$random_digit_or_empty()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_element">}\href{../../charlatan/html/BaseProvider.html#method-random_element}{\code{charlatan::BaseProvider$random_element()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_element_prob">}\href{../../charlatan/html/BaseProvider.html#method-random_element_prob}{\code{charlatan::BaseProvider$random_element_prob()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_int">}\href{../../charlatan/html/BaseProvider.html#method-random_int}{\code{charlatan::BaseProvider$random_int()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="random_letter">}\href{../../charlatan/html/BaseProvider.html#method-random_letter}{\code{charlatan::BaseProvider$random_letter()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="randomize_nb_elements">}\href{../../charlatan/html/BaseProvider.html#method-randomize_nb_elements}{\code{charlatan::BaseProvider$randomize_nb_elements()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-allowed_locales"></a>}}
\if{latex}{\out{\hypertarget{method-allowed_locales}{}}}
\subsection{Method \code{allowed_locales()}}{
fetch the allowed locales for this provider
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$allowed_locales()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{InternetProvider} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$new(locale = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{locale}}{(character) the locale to use. See
\verb{$allowed_locales()} for locales supported (default: en_US)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{InternetProvider} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_ascii"></a>}}
\if{latex}{\out{\hypertarget{method-to_ascii}{}}}
\subsection{Method \code{to_ascii()}}{
convert to ascii
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$to_ascii(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{the stringn to convert to ascii}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-email"></a>}}
\if{latex}{\out{\hypertarget{method-email}{}}}
\subsection{Method \code{email()}}{
get an email address
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$email(domain = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{domain}}{(character) a domain name, if not given, a random
name is chosen}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-safe_email"></a>}}
\if{latex}{\out{\hypertarget{method-safe_email}{}}}
\subsection{Method \code{safe_email()}}{
get a safe email address
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$safe_email()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-free_email"></a>}}
\if{latex}{\out{\hypertarget{method-free_email}{}}}
\subsection{Method \code{free_email()}}{
a free email address
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$free_email()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-company_email"></a>}}
\if{latex}{\out{\hypertarget{method-company_email}{}}}
\subsection{Method \code{company_email()}}{
company email address
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$company_email()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ascii_email"></a>}}
\if{latex}{\out{\hypertarget{method-ascii_email}{}}}
\subsection{Method \code{ascii_email()}}{
ascii email address
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$ascii_email()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ascii_safe_email"></a>}}
\if{latex}{\out{\hypertarget{method-ascii_safe_email}{}}}
\subsection{Method \code{ascii_safe_email()}}{
safe ascii email address
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$ascii_safe_email()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ascii_free_email"></a>}}
\if{latex}{\out{\hypertarget{method-ascii_free_email}{}}}
\subsection{Method \code{ascii_free_email()}}{
an ascii free email address
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$ascii_free_email()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ascii_company_email"></a>}}
\if{latex}{\out{\hypertarget{method-ascii_company_email}{}}}
\subsection{Method \code{ascii_company_email()}}{
ascii company email address
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$ascii_company_email()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-user_name"></a>}}
\if{latex}{\out{\hypertarget{method-user_name}{}}}
\subsection{Method \code{user_name()}}{
a user name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$user_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tld"></a>}}
\if{latex}{\out{\hypertarget{method-tld}{}}}
\subsection{Method \code{tld()}}{
a tld
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$tld()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-free_email_domain"></a>}}
\if{latex}{\out{\hypertarget{method-free_email_domain}{}}}
\subsection{Method \code{free_email_domain()}}{
free email domain name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$free_email_domain()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-url"></a>}}
\if{latex}{\out{\hypertarget{method-url}{}}}
\subsection{Method \code{url()}}{
a url
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$url(schemes = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{schemes}}{(character vector) a url scheme, defaults are http and https}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-domain_name"></a>}}
\if{latex}{\out{\hypertarget{method-domain_name}{}}}
\subsection{Method \code{domain_name()}}{
Produce an Internet domain name with the specified
number of subdomain levels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$domain_name(levels = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{levels}}{(integer) how many levels, must be >1}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-domain_word"></a>}}
\if{latex}{\out{\hypertarget{method-domain_word}{}}}
\subsection{Method \code{domain_word()}}{
a domain word
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$domain_word()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ipv4"></a>}}
\if{latex}{\out{\hypertarget{method-ipv4}{}}}
\subsection{Method \code{ipv4()}}{
a ipv4 address
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$ipv4()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-mac_address"></a>}}
\if{latex}{\out{\hypertarget{method-mac_address}{}}}
\subsection{Method \code{mac_address()}}{
a mac address
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$mac_address()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-uri_page"></a>}}
\if{latex}{\out{\hypertarget{method-uri_page}{}}}
\subsection{Method \code{uri_page()}}{
a uri page
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$uri_page()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-uri_path"></a>}}
\if{latex}{\out{\hypertarget{method-uri_path}{}}}
\subsection{Method \code{uri_path()}}{
a uri path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$uri_path(deep = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{how deep to go, an integer, if not given an integer
between 1 and 4 (inclusive) is chosen}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-uri_extension"></a>}}
\if{latex}{\out{\hypertarget{method-uri_extension}{}}}
\subsection{Method \code{uri_extension()}}{
a uri extension
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$uri_extension()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-uri"></a>}}
\if{latex}{\out{\hypertarget{method-uri}{}}}
\subsection{Method \code{uri()}}{
a uri
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$uri()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-slug"></a>}}
\if{latex}{\out{\hypertarget{method-slug}{}}}
\subsection{Method \code{slug()}}{
a slug
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$slug(value = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{(character) a string, if given, returns itself, if not, uses
\link{LoremProvider} to get a random string. default: \code{NULL}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-image_url"></a>}}
\if{latex}{\out{\hypertarget{method-image_url}{}}}
\subsection{Method \code{image_url()}}{
Returns URL to placeholder image -
Example: http://placehold.it/640x480
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$image_url(width = NULL, height = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{width}}{image width, in pixels}

\item{\code{height}}{image height, in pixels}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InternetProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
