% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\docType{data}
\name{JobProvider}
\alias{JobProvider}
\title{JobProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
JobProvider
}
\description{
JobProvider
}
\examples{
\donttest{
z <- JobProvider$new()
z$render()
}
}
\keyword{internal}

