% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs-provider.R
\docType{data}
\name{JobProvider}
\alias{JobProvider}
\alias{job_provider_locales}
\title{JobProvider}
\format{An object of class \code{character} of length 9.}
\usage{
job_provider_locales
}
\arguments{
\item{locale}{(character) the locale to use. options: en_US (default),
fr_FR, fr_CH, hr_FR, fa_IR, pl_PL, ru_RU, uk_UA, zh_TW.}
}
\description{
JobProvider
}
\details{
\strong{Methods}
\describe{
\item{\code{render()}}{
Make a job
}
}
}
\examples{
z <- JobProvider$new()
z$render()

z <- JobProvider$new(locale = "fr_FR")
z$locale
z$render()

z <- JobProvider$new(locale = "hr_HR")
z$locale
z$render()

z <- JobProvider$new(locale = "fa_IR")
z$locale
z$render()
}
\keyword{datasets}
\keyword{internal}
