% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorem-provider.R
\docType{data}
\name{LoremProvider}
\alias{LoremProvider}
\alias{lorem_provider_locales}
\title{LoremProvider}
\format{An object of class \code{character} of length 2.}
\usage{
lorem_provider_locales
}
\description{
LoremProvider
}
\details{
\strong{Methods}
  \describe{
   \item{\code{word}}{
    get a word
  }
   \item{\code{words}}{
    get words
  }
   \item{\code{sentence}}{
    get a sentence
  }
   \item{\code{sentences}}{
    get sentences
  }
 }
}
\examples{
(x <- LoremProvider$new())
x$locale 
x$word()
x$words(3)
x$words(6)
x$sentence()
x$sentences(3)
x$sentences(6)

(x <- LoremProvider$new(locale = "ar_AA"))
x$word()
}
\keyword{datasets}
\keyword{internal}
