% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-provider.R
\docType{data}
\name{MiscProvider}
\alias{MiscProvider}
\title{MiscProvider}
\description{
MiscProvider
}
\details{
\strong{Methods}
\itemize{
\item \code{language_locale_codes} - list of language locales supported by
Linux Mint
\item \code{locale()} - locale
\item \code{language_code()} - language code
\item \code{boolean()} - a boolean, \code{TRUE} or \code{FALSE}
\item \code{null_boolean()} - a boolean, \code{TRUE} or \code{FALSE}, or NULL
}
}
\examples{
(x <- MiscProvider$new())
x$language_locale_codes
x$language_code()
x$locale()
x$boolean()
x$null_boolean()
}
\keyword{internal}
