% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-provider.R
\name{MiscProvider}
\alias{MiscProvider}
\title{MiscProvider}
\description{
miscellaneous methods
}
\examples{
(x <- MiscProvider$new())
x$language_locale_codes
x$language_code()
x$locale()
x$boolean()
x$null_boolean()
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{language_locale_codes}}{(list) locale codes by locale family}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MiscProvider-boolean}{\code{MiscProvider$boolean()}}
\item \href{#method-MiscProvider-null_boolean}{\code{MiscProvider$null_boolean()}}
\item \href{#method-MiscProvider-locale}{\code{MiscProvider$locale()}}
\item \href{#method-MiscProvider-language_code}{\code{MiscProvider$language_code()}}
\item \href{#method-MiscProvider-clone}{\code{MiscProvider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiscProvider-boolean"></a>}}
\if{latex}{\out{\hypertarget{method-MiscProvider-boolean}{}}}
\subsection{Method \code{boolean()}}{
get a random boolean, \code{TRUE} or \code{FALSE}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiscProvider$boolean(chance_of_getting_true = 50)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{chance_of_getting_true}}{(integer) an integer, default: 50}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiscProvider-null_boolean"></a>}}
\if{latex}{\out{\hypertarget{method-MiscProvider-null_boolean}{}}}
\subsection{Method \code{null_boolean()}}{
get a random boolean, \code{TRUE} or \code{FALSE}, or NULL
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiscProvider$null_boolean()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiscProvider-locale"></a>}}
\if{latex}{\out{\hypertarget{method-MiscProvider-locale}{}}}
\subsection{Method \code{locale()}}{
get a random locale
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiscProvider$locale()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiscProvider-language_code"></a>}}
\if{latex}{\out{\hypertarget{method-MiscProvider-language_code}{}}}
\subsection{Method \code{language_code()}}{
random language code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiscProvider$language_code()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiscProvider-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MiscProvider-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiscProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
