% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-provider.R
\docType{data}
\name{MiscProvider}
\alias{MiscProvider}
\title{MiscProvider}
\description{
MiscProvider
}
\details{
\strong{Methods}
  \describe{
   \item{\code{to_ascii}}{
    convert to ascii
  }
   \item{\code{email}}{
    get an email address
  }
 }
}
\examples{
(x <- MiscProvider$new())
x$language_locale_codes
x$language_code()
x$locale()
}
\keyword{internal}
