% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing-data-provider.R
\docType{data}
\name{MissingDataProvider}
\alias{MissingDataProvider}
\title{MissingDataProvider}
\description{
MissingDataProvider
}
\details{
\strong{Methods}
  \describe{
    \item{\code{make_missing()}}{
      make missing data
      \itemize{
       \item x: a vector of characters, numeric, integers, logicals, etc
      }
    }
  }
}
\examples{
z <- MissingDataProvider$new()
z$make_missing(letters)
}
\keyword{internal}

