% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing-data-provider.R
\docType{data}
\name{MissingDataProvider}
\alias{MissingDataProvider}
\title{MissingDataProvider}
\description{
MissingDataProvider
}
\details{
\strong{Methods}
\describe{
\item{\code{make_missing(x)}}{
make missing data
\itemize{
\item x: a vector of characters, numeric, integers, logicals, etc
}
}
}

This method picks a random number (\code{N}) of slots in the input
vector \code{x} (up to \code{length(x)}). Then picks \code{N} random
positions and replaces them with NA matching the input class.
}
\examples{
z <- MissingDataProvider$new()
z$make_missing(x = letters)
}
\keyword{internal}
