% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numerics-provider.R
\docType{data}
\name{NumericsProvider}
\alias{NumericsProvider}
\title{NumericsProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
NumericsProvider
}
\arguments{
\item{n}{number of values}

\item{mean}{mean value}

\item{sd}{standard deviation}

\item{min}{minimum value}

\item{max}{maximum value}

\item{shape1, shape2}{non-negative parameters of the Beta distribution}

\item{ncp}{non-centrality parameter}
}
\description{
NumericsProvider
}
\examples{
z <- NumericsProvider$new()

z$double()
z$double(10)
z$double(10, mean = 100)
z$double(10, mean = 100, sd = 17)

z$integer()
z$integer(10)
z$integer(10, 1, 20)
z$integer(10, 1, 10000000L)

z$unif()
z$norm()
z$lnorm(10)
z$beta(10, 1, 1)
}
\keyword{internal}

