% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person-provider.R
\name{PersonProvider}
\alias{PersonProvider}
\title{PersonProvider}
\value{
A PersonProvider object that can create names.
}
\description{
PersonProvider

PersonProvider
}
\details{
Methods for Persons, methods for generating names.
}
\note{
You cannot instantiate the Parent providers. You must use one of
the localized one.
}
\seealso{
Other ParentProviders: 
\code{\link{AddressProvider}},
\code{\link{ColorProvider}},
\code{\link{CompanyProvider}},
\code{\link{InternetProvider}},
\code{\link{JobProvider}},
\code{\link{LoremProvider}},
\code{\link{PhoneNumberProvider}},
\code{\link{TaxonomyProvider}},
\code{\link{UserAgentProvider}}
}
\concept{ParentProviders}
\section{Super classes}{
\code{\link[charlatan:BareProvider]{charlatan::BareProvider}} -> \code{\link[charlatan:BaseProvider]{charlatan::BaseProvider}} -> \code{PersonProvider}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{messy}}{show current setting for messy. Either TRUE or FALSE depending
on configuration and if this is even possible for the locale.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PersonProvider-new}{\code{PersonProvider$new()}}
\item \href{#method-PersonProvider-messy_is_possible}{\code{PersonProvider$messy_is_possible()}}
\item \href{#method-PersonProvider-change_messy}{\code{PersonProvider$change_messy()}}
\item \href{#method-PersonProvider-render}{\code{PersonProvider$render()}}
\item \href{#method-PersonProvider-messy_switch}{\code{PersonProvider$messy_switch()}}
\item \href{#method-PersonProvider-first_name}{\code{PersonProvider$first_name()}}
\item \href{#method-PersonProvider-first_name_female}{\code{PersonProvider$first_name_female()}}
\item \href{#method-PersonProvider-first_name_male}{\code{PersonProvider$first_name_male()}}
\item \href{#method-PersonProvider-last_name}{\code{PersonProvider$last_name()}}
\item \href{#method-PersonProvider-last_name_female}{\code{PersonProvider$last_name_female()}}
\item \href{#method-PersonProvider-last_name_male}{\code{PersonProvider$last_name_male()}}
\item \href{#method-PersonProvider-prefix}{\code{PersonProvider$prefix()}}
\item \href{#method-PersonProvider-prefix_female}{\code{PersonProvider$prefix_female()}}
\item \href{#method-PersonProvider-prefix_male}{\code{PersonProvider$prefix_male()}}
\item \href{#method-PersonProvider-suffix}{\code{PersonProvider$suffix()}}
\item \href{#method-PersonProvider-suffix_female}{\code{PersonProvider$suffix_female()}}
\item \href{#method-PersonProvider-suffix_male}{\code{PersonProvider$suffix_male()}}
\item \href{#method-PersonProvider-clone}{\code{PersonProvider$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="bothify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-bothify'><code>charlatan::BareProvider$bothify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="lexify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-lexify'><code>charlatan::BareProvider$lexify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="numerify"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-numerify'><code>charlatan::BareProvider$numerify()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit'><code>charlatan::BareProvider$random_digit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero'><code>charlatan::BareProvider$random_digit_not_zero()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_not_zero_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_not_zero_or_empty'><code>charlatan::BareProvider$random_digit_not_zero_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_digit_or_empty"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_digit_or_empty'><code>charlatan::BareProvider$random_digit_or_empty()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element'><code>charlatan::BareProvider$random_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_element_prob"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_element_prob'><code>charlatan::BareProvider$random_element_prob()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_int"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_int'><code>charlatan::BareProvider$random_int()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="random_letter"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-random_letter'><code>charlatan::BareProvider$random_letter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BareProvider" data-id="randomize_nb_elements"><a href='../../charlatan/html/BareProvider.html#method-BareProvider-randomize_nb_elements'><code>charlatan::BareProvider$randomize_nb_elements()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="allowed_locales"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-allowed_locales'><code>charlatan::BaseProvider$allowed_locales()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="check_locale"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-check_locale'><code>charlatan::BaseProvider$check_locale()</code></a></span></li>
<li><span class="pkg-link" data-pkg="charlatan" data-topic="BaseProvider" data-id="print"><a href='../../charlatan/html/BaseProvider.html#method-BaseProvider-print'><code>charlatan::BaseProvider$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-new"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{PersonProvider} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$new(messy = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{messy}}{make it messy}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-messy_is_possible"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-messy_is_possible}{}}}
\subsection{Method \code{messy_is_possible()}}{
internal function to figure out if
messy is a valid option for this locale.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$messy_is_possible()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-change_messy"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-change_messy}{}}}
\subsection{Method \code{change_messy()}}{
Change messy (if possible)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$change_messy(messy)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{messy}}{TRUE or FALSE}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-render"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-render}{}}}
\subsection{Method \code{render()}}{
Make a person's name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$render(fmt = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fmt}}{(character) a name format, default: \code{NULL}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-messy_switch"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-messy_switch}{}}}
\subsection{Method \code{messy_switch()}}{
messy switch (internal).
Always return a text, when messy is allowed return a messy
version, but otherwise return a clean version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$messy_switch(clean_choice, messy_choice)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{clean_choice}}{the clean version}

\item{\code{messy_choice}}{the messy version}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-first_name"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-first_name}{}}}
\subsection{Method \code{first_name()}}{
make a first name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$first_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-first_name_female"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-first_name_female}{}}}
\subsection{Method \code{first_name_female()}}{
make a female first name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$first_name_female()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-first_name_male"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-first_name_male}{}}}
\subsection{Method \code{first_name_male()}}{
make a male first name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$first_name_male()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-last_name"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-last_name}{}}}
\subsection{Method \code{last_name()}}{
make a last name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$last_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-last_name_female"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-last_name_female}{}}}
\subsection{Method \code{last_name_female()}}{
make a female last name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$last_name_female()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-last_name_male"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-last_name_male}{}}}
\subsection{Method \code{last_name_male()}}{
make a male last name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$last_name_male()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-prefix"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-prefix}{}}}
\subsection{Method \code{prefix()}}{
make a name prefix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$prefix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-prefix_female"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-prefix_female}{}}}
\subsection{Method \code{prefix_female()}}{
make a female name prefix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$prefix_female()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-prefix_male"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-prefix_male}{}}}
\subsection{Method \code{prefix_male()}}{
make a male name prefix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$prefix_male()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-suffix"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-suffix}{}}}
\subsection{Method \code{suffix()}}{
make a name suffix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$suffix()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-suffix_female"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-suffix_female}{}}}
\subsection{Method \code{suffix_female()}}{
make a female name suffix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$suffix_female()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-suffix_male"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-suffix_male}{}}}
\subsection{Method \code{suffix_male()}}{
make a male name suffix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$suffix_male()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PersonProvider-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PersonProvider-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PersonProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
