% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person-provider.R
\docType{data}
\name{PersonProvider}
\alias{PersonProvider}
\title{PersonProvider}
\arguments{
\item{locale}{(character) the locale to use. options: en_US (default),
fr_FR, fr_CH, hr_FR, fa_IR, pl_PL, ru_RU, uk_UA, zh_TW.}
}
\description{
PersonProvider
}
\details{
\strong{Methods}
  \describe{
    \item{\code{render()}}{
      Make a person's name
    }
  }
}
\examples{
x <- PersonProvider$new()
x$render()

z <- PersonProvider$new(locale = "fr_FR")
z$locale
z$render()

PersonProvider$new(locale = "fr_CH")$render()
PersonProvider$new(locale = "fi_FI")$render()
PersonProvider$new(locale = "fa_IR")$render()
PersonProvider$new(locale = "es_ES")$render()
PersonProvider$new(locale = "de_DE")$render()
PersonProvider$new(locale = "de_AT")$render()
PersonProvider$new(locale = "cs_CZ")$render()
PersonProvider$new(locale = "bg_BG")$render()
PersonProvider$new(locale = "dk_DK")$render()
}
\keyword{internal}
