% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person-provider.R
\docType{data}
\name{PersonProvider}
\alias{PersonProvider}
\title{PersonProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PersonProvider
}
\description{
PersonProvider
}
\examples{
\donttest{
x <- PersonProvider$new()
x$render()
}
}
\keyword{internal}

