% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/person-provider.R
\docType{data}
\name{PersonProvider}
\alias{PersonProvider}
\alias{person_provider_locales}
\title{PersonProvider}
\format{An object of class \code{character} of length 12.}
\usage{
person_provider_locales
}
\arguments{
\item{locale}{(character) the locale to use. options: en_US (default),
fr_FR, fr_CH, fi_FI, fa_IR, es_ES, es_MX, de_DE, de_AT, cs_CZ, bg_BG,
dk_DK}

\item{messy}{(logical) make some messy data. Default: \code{FALSE}}
}
\description{
PersonProvider
}
\details{
\strong{Methods}
  \describe{
    \item{\code{render()}}{
      Make a person's name
    }
  }
}
\examples{
x <- PersonProvider$new()
x$render()

z <- PersonProvider$new(locale = "fr_FR")
z$locale
z$render()

z <- PersonProvider$new(locale = "es_MX")
z$locale
z$render()

PersonProvider$new(locale = "fr_CH")$render()
PersonProvider$new(locale = "fi_FI")$render()
PersonProvider$new(locale = "fa_IR")$render()
PersonProvider$new(locale = "es_ES")$render()
PersonProvider$new(locale = "de_DE")$render()
PersonProvider$new(locale = "de_AT")$render()
PersonProvider$new(locale = "cs_CZ")$render()
PersonProvider$new(locale = "bg_BG")$render()
PersonProvider$new(locale = "dk_DK")$render()
}
\keyword{datasets}
\keyword{internal}
