% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonenumbers-provider.R
\docType{data}
\name{PhoneNumberProvider}
\alias{PhoneNumberProvider}
\title{PhoneNumberProvider}
\description{
PhoneNumberProvider
}
\details{
\strong{Methods}
  \describe{
    \item{\code{render()}}{
      Make a phone number
    }
  }
}
\examples{
z <- PhoneNumberProvider$new()
z$render()
}
\keyword{internal}
