% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonenumbers-provider.R
\docType{data}
\name{PhoneNumberProvider}
\alias{PhoneNumberProvider}
\title{PhoneNumberProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PhoneNumberProvider
}
\description{
PhoneNumberProvider
}
\examples{
\donttest{
z <- PhoneNumberProvider$new()
z$render()
}
}
\keyword{internal}

