% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence-provider.R
\name{SequenceProvider}
\alias{SequenceProvider}
\title{SequenceProvider}
\description{
SequenceProvider

SequenceProvider
}
\details{
\strong{Methods}
\itemize{
\item \code{render(length)} - Make a sequence
\itemize{
\item length: sequence length. default: 30
}
}
}
\examples{
z <- SequenceProvider$new()
z$render()
z$render(10)
z$render(100)
z$render(500)
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-render}{\code{SequenceProvider$render()}}
\item \href{#method-clone}{\code{SequenceProvider$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-render"></a>}}
\subsection{Method \code{render()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequenceProvider$render(length = 30)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{length}}{(integer) length of sequence to create}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SequenceProvider$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
