% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence-provider.R
\docType{data}
\name{SequenceProvider}
\alias{SequenceProvider}
\title{SequenceProvider}
\arguments{
\item{length}{(integer) length of sequence to create}
}
\description{
SequenceProvider
}
\details{
\strong{Methods}
  \describe{
    \item{\code{render()}}{
      Make a sequence
      \itemize{
       \item length: sequence length. default: 30
      }
    }
  }
}
\examples{
z <- SequenceProvider$new()
z$render()
z$render(10)
z$render(100)
z$render(500)
}
\keyword{internal}

