% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy}
\alias{ch_taxonomic_epithet}
\alias{ch_taxonomic_genus}
\alias{ch_taxonomic_species}
\alias{taxonomy}
\title{Create fake taxonomic names}
\usage{
ch_taxonomic_genus(n = 1)

ch_taxonomic_epithet(n = 1)

ch_taxonomic_species(n = 1)
}
\arguments{
\item{n}{(integer) number of things to get, 1 to inifinity}
}
\description{
Create fake taxonomic names
}
\examples{
ch_taxonomic_genus()
ch_taxonomic_genus(10)
ch_taxonomic_genus(500)

ch_taxonomic_epithet()
ch_taxonomic_epithet(10)
ch_taxonomic_epithet(500)

ch_taxonomic_species()
ch_taxonomic_species(10)
ch_taxonomic_species(500)
}

