% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy-provider.R
\docType{data}
\name{TaxonomyProvider}
\alias{TaxonomyProvider}
\title{TaxonomyProvider}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TaxonomyProvider
}
\description{
TaxonomyProvider
}
\details{
names taken from Theplantlist. 500 genera names and 500
epithets were chosen at random from the set of 10,000 names in the
dataset in the \code{taxize} package. Theplantlist is, as it says on the
tin, composed of plant names - so these fake names are derived from
plant names if that matters to you.
}
\section{Taxonomic authority}{

Randomly, the taxonomic authority is in parentheses - which represents
that the given authority was not the original authority.
}
\examples{
(z <- TaxonomyProvider$new())
z$genus()
z$epithet()
z$species()
z$species(authority = TRUE)
z$species(authority = TRUE, date = TRUE)
}
\keyword{internal}

