% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useragent-provider.R
\docType{data}
\name{UserAgentProvider}
\alias{UserAgentProvider}
\alias{useragent_provider_locales}
\title{UserAgentProvider}
\format{An object of class \code{character} of length 1.}
\usage{
useragent_provider_locales
}
\arguments{
\item{locale}{(character) the locale to use. See
\code{useragent_provider_locales} for locales supported (default: en_US)}
}
\description{
UserAgentProvider
}
\details{
\strong{Methods}
\itemize{
\item \code{mac_processor()} - a mac processor
\item \code{linux_processor()} - a linux processor
\item \code{user_agent()} - a user agent string, randomly picks one
from chrome, firefix, safari, opera or internet explorer
\item \code{chrome(version_from, version_to, build_from, build_to)} - a chrome
user agent string
\item \code{firefox()} - a firefox user agent string
\item \code{safari()} - a safari user agent string
\item \code{opera()} - an opera user agent string
\item \code{internet_explorer()} - an internet explorer user agent string
\item \code{windows_platform_token()} - a windows platform token
\item \code{linux_platform_token()} - a linux platform token
\item \code{mac_platform_token()} - a mac platform token
}
}
\examples{
(x <- UserAgentProvider$new())
x$locale
x$mac_processor()
x$linux_processor()
x$user_agent()
x$chrome()
x$firefox()
x$internet_explorer()
x$opera()
x$safari()
}
\keyword{datasets}
\keyword{internal}
